fix: split out US and Global APIs
Split out API endpoint functions for US and Global, creating two new files:
 -soilIDModel_US_v3.py
-soilidModel_Global_v3.py
