+++ python_code/soilIDModel_US_v3.py
@@ -1,6 +1,5 @@
 # Standard libraries
 import collections
-import csv
 import json
 import re
 
@@ -9,8 +8,10 @@ import geopandas as gpd
 import numpy as np
 import pandas as pd
 import requests
-import scipy.stats
 import shapely
+import io
+from scipy.stats import spearmanr
+from skbio.stats.composition import ilr, ilr_inv
 
 # Flask
 from flask import current_app
@@ -18,40 +19,33 @@ from flask import current_app
 # Import local fucntions
 from model.local_functions_SoilID_v3 import (
     agg_data_layer,
-    assign_max_distance_scores,
-    calculate_deltaE2000,
-    calculate_location_score,
-    compute_data_completeness,
     drop_cokey_horz,
     extract_muhorzdata_STATSGO,
     extract_statsgo_mucompdata,
-    extract_values,
-    extract_WISE_data,
     fill_missing_comppct_r,
-    get_WRB_descriptions,
     getCF_fromClass,
     getClay,
-    getDataStore_Connection,
     getOSDCF,
     getProfile,
-    getProfile_SG,
     getProfileLAB,
     getSand,
-    getSG_descriptions,
     getTexture,
     gower_distances,
     lab2munsell,
     load_model_output,
     munsell2rgb,
-    pedon_color,
     process_distance_scores,
     pt2polyDist,
     save_model_output,
     save_rank_output,
-    save_soilgrids_output,
     sda_return,
-    silt_calc,
     trim_fraction,
+    infill_soil_data,
+    slice_and_aggregate_soil_data,
+    acomp,
+    simulate_correlated_triangular,
+    process_data_with_rosetta,
+    aggregate_data,
 )
 from osgeo import ogr
 from pandas.io.json import json_normalize
@@ -70,19 +64,21 @@ from shapely.geometry import Point
 # after user has collected data, call rankPredictionUS/rankPredictionGlobal.
 
 
-#####################################################################################################
-#                                       Database and API Functions                                  #
-#####################################################################################################
+##############################################################################################
+#                                   Database and API Functions                               #
+##############################################################################################
 def findSoilLocation(lon, lat):
     """
-    Determines the location type (US, Global, or None) of the given longitude and latitude based on soil datasets.
+    Determines the location type (US, Global, or None) of the given longitude
+        and latitude based on soil datasets.
 
     Args:
     - lon (float): Longitude of the point.
     - lat (float): Latitude of the point.
 
     Returns:
-    - str or None: 'US' if point is in US soil dataset, 'Global' if in global dataset, None otherwise.
+    - str or None: 'US' if point is in US soil dataset, 'Global' if in global dataset,
+        None otherwise.
     """
 
     drv_h = ogr.GetDriverByName("ESRI Shapefile")
@@ -121,1159 +117,34 @@ def findSoilLocation(lon, lat):
         return "Global"
 
 
-#####################################################################################################
-#                                 getSoilLocationBasedGlobal                                        #
-#####################################################################################################
-def getSoilLocationBasedGlobal(lon, lat, plot_id):
-    # Extract HWSD-WISE Data
-    # Note: Need to convert HWSD shp to gpkg file
-    wise_data = extract_WISE_data(
-        lon,
-        lat,
-        file_path="%s/wise30sec_poly_simp_soil.shp" % current_app.config["DATA_BACKEND"],
-        layer_name=None,
-        buffer_size=0.5,
-    )
-
-    # Component Data
-    mucompdata_pd = wise_data[["MUGLB_NEW", "SU_name", "distance", "PROP", "COMPID", "FAO_SYS"]]
-    mucompdata_pd.columns = ["mukey", "compname", "distance", "share", "cokey", "fss"]
-    mucompdata_pd["distance"] = pd.to_numeric(mucompdata_pd["distance"])
-    mucompdata_pd["share"] = pd.to_numeric(mucompdata_pd["share"])
-    mucompdata_pd = mucompdata_pd.drop_duplicates().reset_index(drop=True)
-
-    # --------------------------------------------------------------------------------------------------------------------------------------------------------
-    #########################################################################################################################################
-    # Individual probability
-    # Based on Fan et al 2018 EQ 1, the conditional probability for each component is calculated by taking the sum of all occurances of a component
-    # in the home and adjacent mapunits and dividing this by the sum of all map units and components. We have modified this approach so that each
-    # instance of a component occurance is evaluated separately and assinged a weight and the max distance score for each component group is assigned
-    # to all component instances.
-    #########################################################################################################################################
-    ExpCoeff = -0.00036888  # Decays to 0.25 @ 10km
-    loc_scores = []
-    mucompdata_grouped = mucompdata_pd.groupby(["mukey", "cokey"], sort=False)
-
-    for (mukey, cokey), group in mucompdata_grouped:
-        loc_score = calculate_location_score(group, ExpCoeff)
-        loc_scores.append({"Cokey": cokey, "Mukey": mukey, "distance_score": round(loc_score, 3)})
-
-    loc_top_pd = pd.DataFrame(loc_scores)
-    loc_top_comp_prob = loc_top_pd.groupby("Cokey").distance_score.sum()
-    loc_bot_prob_sum = loc_top_pd.distance_score.sum()
-    cond_prob = (loc_top_comp_prob / loc_bot_prob_sum).reset_index(name="distance_score")
-
-    mucompdata_pd = pd.merge(mucompdata_pd, cond_prob, on="cokey", how="left")
-    mucompdata_pd = mucompdata_pd.sort_values("distance_score", ascending=False)
-    mucompdata_pd["distance_score_norm"] = (
-        mucompdata_pd.distance_score / mucompdata_pd.distance_score.max()
-    ) * 0.25
-    mucompdata_pd = mucompdata_pd.reset_index(drop=True)
-    mucompdata_pd["distance"] = mucompdata_pd["distance"].round(4)
-    mucompdata_pd["Index"] = mucompdata_pd.index
-
-    # Group by component name
-    mucompdata_grouped = mucompdata_pd.groupby("compname", sort=False)
-
-    # Take at most 12 groups
-    mucompdata_comp_grps = [group for _, group in mucompdata_grouped][:12]
-
-    # Assign max distance scores to all members within each group
-    soilIDList_out = [assign_max_distance_scores(group) for group in mucompdata_comp_grps]
-
-    mucompdata_pd = pd.concat(soilIDList_out).reset_index(drop=True)
-    comp_key = mucompdata_pd["cokey"].tolist()
-
-    # -----------------------------------------------------------------------------------------------------------------
-    # Create horizon data table
-    columns_to_select = [
-        "COMPID",
-        "TopDep",
-        "BotDep",
-        "id",
-        "Layer",
-        "SDTO",
-        "STPC",
-        "CLPC",
-        "CFRAG",
-        "CECS",
-        "PHAQ",
-        "ELCO",
-        "PROP",
-        "SU_name",
-        "FAO_SYS",
-    ]
-    new_column_names = [
-        "cokey",
-        "hzdept_r",
-        "hzdepb_r",
-        "chkey",
-        "hzname",
-        "sandtotal_r",
-        "silttotal_r",
-        "claytotal_r",
-        "total_frag_volume",
-        "CEC",
-        "pH",
-        "EC",
-        "comppct_r",
-        "compname",
-        "fss",
-    ]
-
-    muhorzdata_pd = wise_data[columns_to_select]
-    muhorzdata_pd.columns = new_column_names
-    muhorzdata_pd = muhorzdata_pd[muhorzdata_pd["cokey"].isin(comp_key)]
-    muhorzdata_pd[["hzdept_r", "hzdepb_r"]] = (
-        muhorzdata_pd[["hzdept_r", "hzdepb_r"]].fillna(0).astype(int)
-    )
-    muhorzdata_pd["texture"] = muhorzdata_pd.apply(getTexture, axis=1)
-
-    # Rank components and sort by rank and depth
-    cokey_Index = {key: rank for rank, key in enumerate(comp_key)}
-    muhorzdata_pd["Comp_Rank"] = muhorzdata_pd["cokey"].map(cokey_Index)
-    muhorzdata_pd.sort_values(["Comp_Rank", "hzdept_r"], inplace=True)
-    muhorzdata_pd.drop(columns="Comp_Rank", inplace=True)
-    muhorzdata_pd = muhorzdata_pd.drop_duplicates().reset_index(drop=True)
-
-    # Check for duplicate component instances
-    hz_drop = drop_cokey_horz(muhorzdata_pd)
-    if hz_drop is not None:
-        muhorzdata_pd = muhorzdata_pd[~muhorzdata_pd.cokey.isin(hz_drop)]
-
-    # Update comp_key
-    comp_key = muhorzdata_pd["cokey"].unique().tolist()
-
-    # Subset mucompdata_pd by new comp_key and add suffix to name if there are duplicates
-    mucompdata_pd = mucompdata_pd[mucompdata_pd["cokey"].isin(comp_key)]
-    mucompdata_pd.sort_values(["distance_score", "distance"], ascending=[False, True], inplace=True)
-    mucompdata_pd.reset_index(drop=True, inplace=True)
-
-    # Add suffix to duplicate names
-    name_counts = collections.Counter(mucompdata_pd["compname"])
-    for name, count in name_counts.items():
-        if count > 1:
-            for suffix in range(1, count + 1):
-                mucompdata_pd.loc[mucompdata_pd["compname"] == name, "compname"] = name + str(
-                    suffix
-                )
-
-    # Add modified compname to muhorzdata
-    muhorzdata_name = muhorzdata_pd[["cokey"]].merge(
-        mucompdata_pd[["cokey", "compname"]], on="cokey"
-    )
-    muhorzdata_pd["compname"] = muhorzdata_name["compname"]
-
-    # Group data by cokey for texture
-    muhorzdata_group_cokey = list(muhorzdata_pd.groupby("cokey", sort=False))
-
-    # Initialize lists for storing data
-    getProfile_cokey = []
-    c_bottom_depths = []
-    clay_texture = []
-    snd_lyrs = []
-    cly_lyrs = []
-    txt_lyrs = []
-    hz_lyrs = []
-    rf_lyrs = []
-    cec_lyrs = []
-    ph_lyrs = []
-    ec_lyrs = []
-
-    for group in muhorzdata_group_cokey:
-        profile = (
-            group.sort_values(by="hzdept_r").drop_duplicates(keep="first").reset_index(drop=True)
-        )
-
-        c_very_bottom, sand_pct_intpl = getProfile(profile, "sandtotal_r", c_bot=True)
-        sand_pct_intpl.columns = ["c_sandpct_intpl", "c_sandpct_intpl_grp"]
-
-        def process_profile(param):
-            result = getProfile(profile, param)
-            result.columns = [f"c_{param}_intpl", f"c_{param}_intpl_grp"]
-            return result
-
-        clay_pct_intpl = process_profile("claytotal_r")
-        cf_pct_intpl = process_profile("total_frag_volume")
-        cec_intpl = process_profile("CEC")
-        ph_intpl = process_profile("pH")
-        ec_intpl = process_profile("EC")
-
-        combined_data = pd.concat(
-            [
-                sand_pct_intpl[["c_sandpct_intpl_grp"]],
-                clay_pct_intpl[["c_claypct_intpl_grp"]],
-                cf_pct_intpl[["c_cfpct_intpl_grp"]],
-                profile.compname.unique(),
-                profile.cokey.unique(),
-                profile.comppct_r.unique(),
-            ],
-            axis=1,
-        )
-
-        combined_data.columns = [
-            "sandpct_intpl",
-            "claypct_intpl",
-            "rfv_intpl",
-            "compname",
-            "cokey",
-            "comppct",
-        ]
-
-        c_bottom_temp = pd.DataFrame(
-            {
-                "cokey": [combined_data["cokey"].iloc[0]],
-                "compname": [combined_data["compname"].iloc[0]],
-                "c_very_bottom": [int(c_very_bottom)],
-            }
-        )
-
-        def aggregated_data_layer(data, column_name):
-            return agg_data_layer(
-                data=data[column_name], bottom=c_bottom_temp["c_very_bottom"].iloc[0], depth=True
-            )
-
-        snd_d, hz_depb = aggregated_data_layer(sand_pct_intpl, "c_sandpct_intpl")
-        cly_d = aggregated_data_layer(clay_pct_intpl, "c_claypct_intpl")
-        txt_d = [
-            getTexture(row=None, sand=s, silt=(100 - (s + c)), clay=c) for s, c in zip(snd_d, cly_d)
-        ]
-        txt_d = pd.Series(txt_d, index=snd_d.index)
-
-        rf_d = aggregated_data_layer(cf_pct_intpl, "c_cfpct_intpl_grp")
-        cec_d = aggregated_data_layer(cec_intpl, "c_cec_intpl")
-        ph_d = aggregated_data_layer(ph_intpl, "c_ph_intpl")
-        ec_d = aggregated_data_layer(ec_intpl, "c_ec_intpl")
-
-        # Fill NaN values and append to lists
-        for data_list, data in zip(
-            [snd_lyrs, cly_lyrs, txt_lyrs, rf_lyrs, cec_lyrs, ph_lyrs, ec_lyrs, hz_lyrs],
-            [snd_d, cly_d, txt_d, rf_d, cec_d, ph_d, ec_d, hz_depb],
-        ):
-            data_list.append(dict(data.fillna("")))
-
-        c_bottom_depths.append(c_bottom_temp)
-        getProfile_cokey.append(combined_data)
-
-        comp_texture_list = [x for x in profile.texture.str.lower() if x]
-        clay_val = "Yes" if any("clay" in string for string in comp_texture_list) else "No"
-        clay_texture_temp = pd.DataFrame(
-            {"compname": [combined_data["compname"].iloc[0]], "clay": [clay_val]}
-        )
-        clay_texture.append(clay_texture_temp)
-
-    # Concatenate lists to form DataFrames
-    c_bottom_depths = pd.concat(c_bottom_depths, axis=0)
-    clay_texture = pd.concat(clay_texture, axis=0)
-
-    # Subset mucompdata and muhorzdata DataFrames
-    mucompdata_pd = mucompdata_pd[mucompdata_pd["cokey"].isin(c_bottom_depths.cokey)]
-    muhorzdata_pd = muhorzdata_pd[muhorzdata_pd["cokey"].isin(c_bottom_depths.cokey)]
-
-    # Merge c_bottom_depth and clay_texture with mucompdata
-    mucompdata_pd = pd.merge(
-        mucompdata_pd, c_bottom_depths[["compname", "c_very_bottom"]], on="compname", how="left"
-    )
-    mucompdata_pd = pd.merge(mucompdata_pd, clay_texture, on="compname", how="left")
-
-    # Create index for component instance display
-    mucompdata_comp_grps_list = []
-    mucompdata_comp_grps = [group for _, group in mucompdata_pd.groupby("compname_grp", sort=False)]
-
-    for group in mucompdata_comp_grps:
-        group = group.sort_values("distance").reset_index(drop=True)
-        soilID_rank = [True if idx == 0 else False for idx in range(len(group))]
-
-        group["soilID_rank"] = soilID_rank
-        group["min_dist"] = group.distance.iloc[0]
-
-        mucompdata_comp_grps_list.append(group)
-
-    mucompdata_pd = pd.concat(mucompdata_comp_grps_list).reset_index(drop=True)
-
-    # --------------------------------------------------------------------------
-    # SoilIDList output
-
-    # ------------------------------------------------------------
-    # Format output data
-    soilIDRank_output = [
-        group[["compname", "sandpct_intpl", "claypct_intpl", "rfv_intpl"]]
-        for group in getProfile_cokey
-    ]
-    soilIDRank_output_pd = pd.concat(soilIDRank_output, axis=0).reset_index(drop=True)
-
-    mucompdata_cond_prob = mucompdata_pd.sort_values(
-        "distance_score_norm", ascending=False
-    ).reset_index(drop=True)
-
-    # Determine rank location
-    rank_id = 1
-    Rank_Loc = []
-    for rank in mucompdata_cond_prob["soilID_rank"]:
-        Rank_Loc.append(str(rank_id) if rank else "Not Displayed")
-        rank_id += rank  # Increase rank_id only if rank is True
-    mucompdata_cond_prob["Rank_Loc"] = Rank_Loc
-
-    # Handle NaN values
-    mucompdata_cond_prob.replace({np.nan: "", "nan": "", "None": "", None: ""}, inplace=True)
-
-    # Merge component descriptions
-    WRB_Comp_Desc = get_WRB_descriptions(
-        mucompdata_cond_prob["compname_grp"].drop_duplicates().tolist()
-    )
-    mucompdata_cond_prob = pd.merge(
-        mucompdata_cond_prob, WRB_Comp_Desc, left_on="compname_grp", right_on="WRB_tax", how="left"
-    )
-
-    mucomp_index = mucompdata_cond_prob.sort_values(
-        ["soilID_rank", "distance_score_norm"], ascending=[False, False]
-    ).index
-
-    # Extract site information
-    Site = [
-        {
-            "siteData": {
-                "mapunitID": row.mukey,
-                "componentID": row.cokey,
-                "fao": row.fss,
-                "share": row.share,
-                "distance": round(row.distance, 3),
-                "minCompDistance": row.min_dist,
-                "soilDepth": row.c_very_bottom,
-            },
-            "siteDescription": {
-                key: row[key]
-                for key in [
-                    "Description_en",
-                    "Management_en",
-                    "Description_es",
-                    "Management_es",
-                    "Description_ks",
-                    "Management_ks",
-                    "Description_fr",
-                    "Management_fr",
-                ]
-            },
-        }
-        for _, row in mucompdata_cond_prob.iterrows()
-    ]
-
-    # Reorder lists based on mucomp_index
-    hz_lyrs = [hz_lyrs[i] for i in mucomp_index]
-    snd_lyrs = [snd_lyrs[i] for i in mucomp_index]
-    cly_lyrs = [cly_lyrs[i] for i in mucomp_index]
-    txt_lyrs = [txt_lyrs[i] for i in mucomp_index]
-    rf_lyrs = [rf_lyrs[i] for i in mucomp_index]
-    cec_lyrs = [cec_lyrs[i] for i in mucomp_index]
-    ph_lyrs = [ph_lyrs[i] for i in mucomp_index]
-    ec_lyrs = [ec_lyrs[i] for i in mucomp_index]
-
-    output_SoilList = [
-        dict(
-            zip(
-                [
-                    "id",
-                    "site",
-                    "bottom_depth",
-                    "sand",
-                    "clay",
-                    "texture",
-                    "rock_fragments",
-                    "cec",
-                    "ph",
-                    "ec",
-                ],
-                item,
-            )
-        )
-        for item in zip(
-            ID, Site, hz_lyrs, snd_lyrs, cly_lyrs, txt_lyrs, rf_lyrs, cec_lyrs, ph_lyrs, ec_lyrs
-        )
-    ]
-
-    # Save data
-    if plot_id is None:
-        soilIDRank_output_pd.to_csv(
-            "%s/soilIDRank_ofile1.csv" % current_app.config["DATA_BACKEND"], index=None, header=True
-        )
-        mucompdata_cond_prob.to_csv(
-            "%s/soilIDRank_ofile2.csv" % current_app.config["DATA_BACKEND"], index=None, header=True
-        )
-    else:
-        save_model_output(
-            plot_id,
-            model_version,
-            json.dumps(
-                {
-                    "metadata": {
-                        "location": "global",
-                        "model": "v2",
-                        "unit_measure": {
-                            "distance": "m",
-                            "depth": "cm",
-                            "cec": "cmol(c)/kg",
-                            "clay": "%",
-                            "rock_fragments": "cm3/100cm3",
-                            "sand": "%",
-                            "ec": "ds/m",
-                        },
-                    },
-                    "soilList": output_SoilList,
-                }
-            ),
-            soilIDRank_output_pd.to_csv(index=None, header=True),
-            mucompdata_cond_prob.to_csv(index=None, header=True),
-        )
-
-    # Return the JSON output
-    return {
-        "metadata": {
-            "location": "global",
-            "model": "v2",
-            "unit_measure": {
-                "distance": "m",
-                "depth": "cm",
-                "cec": "cmol(c)/kg",
-                "clay": "%",
-                "rock_fragments": "cm3/100cm3",
-                "sand": "%",
-                "ec": "ds/m",
-            },
-        },
-        "soilList": output_SoilList,
-    }
-
-
-#############################################################################################
-#                                   rankPredictionGlobal                                    #
-#############################################################################################
-def rankPredictionGlobal(
-    lon, lat, soilHorizon, horizonDepth, rfvDepth, lab_Color, bedrock, cracks, plot_id=None
-):
-    # ---------------------------------------------------------------------------------------
-    # ------ Load in user data --------#
-
-    # Create the dataframe
-    soil_df = pd.DataFrame(
-        {
-            "soilHorizon": soilHorizon,
-            "horizonDepth": horizonDepth,
-            "rfvDepth": rfvDepth,
-            "lab_Color": lab_Color,
-        }
-    )
-
-    # Cleanup the dataframe
-    soil_df.dropna(how="all", inplace=True)
-    soil_df["bottom"] = soil_df["horizonDepth"]
-
-    # Replace NaN values with None for consistency
-    soil_df = soil_df.where(pd.notna(soil_df), None)
-
-    # Compute the 'top' column
-    top = [0] + [int(depth) for depth in soil_df["horizonDepth"][:-1]]
-    soil_df["top"] = top
-
-    # Set 'horizonDepth' as the index
-    soil_df.set_index("horizonDepth", inplace=True)
-
-    # Drop rows where 'soilHorizon' and 'rfvDepth' are both missing
-    soil_df_slice = soil_df.dropna(subset=["soilHorizon", "rfvDepth"], how="all")
-
-    # If the slice is not empty, reset its index
-    if not soil_df_slice.empty:
-        soil_df_slice.reset_index(inplace=True)
-        soil_df_slice.rename(columns={"horizonDepth": "index"}, inplace=True)
-        soil_df_slice.set_index("index", inplace=True)
-
-    if soil_df_slice is not None:
-        # If bedrock has been recorded and the lowest soil depth associated with data 
-        # is greater than bedrock, then change lowest soil depth to bedrock depth
-        if bedrock and soil_df_slice["bottom"].iloc[-1] > bedrock:
-            soil_df_slice["bottom"].iloc[-1] = bedrock
-            soil_df["bottom"][soil_df_slice.index[-1]] = bedrock
-            soil_df.reset_index(drop=True, inplace=True)
-
-            # Infill missing horizons
-            for j in range(len(soil_df) - 1):
-                if soil_df["top"].iloc[j + 1] > soil_df["bottom"].iloc[j]:
-                    layer_add = pd.DataFrame(
-                        {
-                            "soilHorizon": [None],
-                            "rfvDepth": [None],
-                            "lab_Color": [None],
-                            "bottom": [soil_df["top"].iloc[j + 1]],
-                            "top": [soil_df["bottom"].iloc[j]],
-                        }
-                    )
-                    soil_df = (
-                        pd.concat([soil_df, layer_add], axis=0)
-                        .sort_values("top")
-                        .reset_index(drop=True)
-                    )
-
-            soil_df.where(pd.notnull(soil_df), None, inplace=True)
-
-        # Create index list of soil slices where user data exists
-        soil_df_slice.reset_index(drop=True, inplace=True)
-        pedon_slice_index = [
-            j
-            for i in range(len(soil_df_slice))
-            for j in range(int(soil_df_slice["top"].iloc[i]), int(soil_df_slice["bottom"].iloc[i]))
-            if j < 120
-        ]
-        if bedrock:
-            pedon_slice_index.extend(range(bedrock, 120))
-
-        # Soil properties to lists
-        soilHorizon, rfvDepth = soil_df["soilHorizon"].tolist(), soil_df["rfvDepth"].tolist()
-        horizonDepthB, horizonDepthT = list(map(int, soil_df["bottom"])), list(
-            map(int, soil_df["top"])
-        )
-        lab_Color = soil_df["lab_Color"]
-
-        # Format and generate user data
-        spt = [getSand(sh) for sh in soilHorizon]
-        cpt = [getClay(sh) for sh in soilHorizon]
-        p_cfg = [getCF_fromClass(rf) for rf in rfvDepth]
-
-        p_sandpct_intpl = [
-            spt[i]
-            for i in range(len(soilHorizon))
-            for _ in range(horizonDepthT[i], horizonDepthB[i])
-        ]
-        p_claypct_intpl = [
-            cpt[i]
-            for i in range(len(soilHorizon))
-            for _ in range(horizonDepthT[i], horizonDepthB[i])
-        ]
-        p_cfg_intpl = [
-            p_cfg[i]
-            for i in range(len(soilHorizon))
-            for _ in range(horizonDepthT[i], horizonDepthB[i])
-        ]
-
-        # Adjust depth interval of user data
-        for data_list in [p_sandpct_intpl, p_claypct_intpl, p_cfg_intpl]:
-            if len(data_list) > 120:
-                data_list = data_list[:120]
-            else:
-                data_list.extend([np.nan] * (120 - len(data_list)))
-
-        p_hz_data = pd.DataFrame(
-            {
-                "compname": "sample_pedon",
-                "sandpct_intpl": p_sandpct_intpl,
-                "claypct_intpl": p_claypct_intpl,
-                "rfv_intpl": p_cfg_intpl,
-            }
-        )
-
-        # Drop empty data columns and depth slices
-        p_hz_data.dropna(axis=1, how="all", inplace=True)
-        p_hz_data = p_hz_data[p_hz_data.index.isin(pedon_slice_index)]
-        p_hz_data_names = p_hz_data.columns.tolist()
-
-    else:
-        p_hz_data, p_sandpct_intpl, p_claypct_intpl, p_cfg_intpl = None, [], [], []
-        p_bottom_depth = pd.DataFrame(
-            {"cokey": [-999], "compname": ["sample_pedon"], "bottom_depth": [0]}
-        )
-
-    lab_Color = pd.Series(lab_Color)
-    if not lab_Color.isnull().all():
-        lab_Color = [[np.nan, np.nan, np.nan] if x is None else x for x in lab_Color]
-        lab_Color = pd.DataFrame(lab_Color)
-        pedon_LAB = pedon_color(lab_Color, horizonDepth)
-
-        if not np.isnan(pedon_LAB).all():
-            refs = {
-                "white": [100.0, 0.0, 0.0],
-                "red": [53.2, 80.1, 67.2],
-                "yellow": [97.1, -21.6, 94.5],
-            }
-
-            cr_df = [calculate_deltaE2000(pedon_LAB, refs[color]) for color in refs]
-            cr_df = pd.Series(cr_df)
-        else:
-            cr_df = pd.Series([np.nan])
-    else:
-        cr_df = pd.Series([np.nan])
-
-    # Calculate Data Completeness Score
-    data_completeness, text_completeness = compute_data_completeness(
-        bedrock, p_sandpct_intpl, soilHorizon, p_cfg_intpl, rfvDepth, cracks, cr_df
-    )
-
-    # ---------------------------------------------------------------------------------
-    # Load in component data from soilIDList
-    # Initialize
-    record_id = None
-
-    # If no plot_id is provided, load data from file
-    if plot_id is None:
-        soilIDRank_output = pd.read_csv(
-            "{}/soilIDRank_ofile1.csv".format(current_app.config["DATA_BACKEND"])
-        )
-        mucompdata = pd.read_csv(
-            "{}/soilIDRank_ofile2.csv".format(current_app.config["DATA_BACKEND"])
-        )
-
-    # If plot_id is provided, load data from the database
-    else:
-        modelRun = load_model_output(plot_id)
-
-        # Check if modelRun data was successfully fetched
-        if modelRun:
-            record_id = modelRun[0]
-            soilIDRank_output = pd.read_csv(io.StringIO(modelRun[2]))
-            mucompdata = pd.read_csv(io.StringIO(modelRun[3]))
-        else:
-            return "Cannot find a plot with this ID"
-
-    # Group the soilIDRank_output dataframe by 'compname' and return
-    grouped_soil_data = [group for _, group in soilIDRank_output.groupby("compname", sort=False)]
-
-    # Create soil depth DataFrame and subset component depths based on max 
-    # user depth if no bedrock specified
-    c_bottom_depths = mucompdata_pd[["compname", "c_very_bottom"]].rename(
-        columns={"c_very_bottom": "bottom_depth"}
-    )
-    slices_of_soil = pd.concat([p_bottom_depth, c_bottom_depths], axis=0).reset_index(drop=True)
-
-    compnames = mucompdata_pd[["compname", "compname_grp"]]
-
-    # If bedrock is not specified, determine max_depth based on user recorded 
-    # depth (limited to 120 cm)
-    max_depth = (
-        120
-        if bedrock is None and p_bottom_depth.bottom_depth.values[0] > 120
-        else p_bottom_depth.bottom_depth.values[0]
-    )
-
-    # If bedrock is specified, truncate to the app's max depth of 120 cm
-    max_depth = 120 if bedrock is not None else max_depth
-
-    # Adjust slices of soil if they exceed the determined max_depth
-    slices_of_soil.loc[slices_of_soil.bottom_depth > max_depth, "bottom_depth"] = max_depth
-    slices_of_non_soil = max_depth - slices_of_soil.bottom_depth
-
-    # Generate a matrix describing soil (1) vs. non-soil (0) for each slice
-    soil_matrix = pd.DataFrame(
-        np.nan, index=np.arange(max_depth), columns=np.arange(len(slices_of_soil))
-    )
-    for i in range(len(slices_of_soil)):
-        slice_end = slices_of_soil.bottom_depth.iloc[i]
-        soil_matrix.iloc[:slice_end, i] = 1
-        soil_matrix.iloc[slice_end:max_depth, i] = 0
-
-    # Determine if user has entered horizon data and if so, subset component horizon 
-    # data based on user input data
-    if p_hz_data:
-        # Subset component soil properties to match user measured properties
-        soilIDRank_output_pd = soilIDRank_output_pd[p_hz_data_names]
-        # Subset soil_matrix to user measured slices
-        soil_matrix = soil_matrix.loc[pedon_slice_index]
-
-        # Horizon Data Similarity
-        if p_bottom_depth.bottom_depth.any() > 0:
-            horz_vars_group_cokey = [
-                group for _, group in soilIDRank_output_pd.groupby("compname", sort=False)
-            ]
-
-            # Create lists to store component statuses
-            Comp_Rank_Status, Comp_Missing_Status, Comp_name = [], [], []
-
-            # Check component status
-            for group in horz_vars_group_cokey:
-                subset_group = group[p_hz_data_names].drop(columns="compname")
-                if subset_group.isnull().values.all():
-                    Comp_Rank_Status.append("Not Ranked")
-                    Comp_Missing_Status.append("No Data")
-                else:
-                    Comp_Rank_Status.append("Ranked")
-                    Comp_Missing_Status.append(
-                        "Missing Data" if subset_group.isnull().values.any() else "Data Complete"
-                    )
-                Comp_name.append(group["compname"].unique()[0])
-
-            Rank_Filter = pd.DataFrame(
-                {
-                    "compname": Comp_name,
-                    "rank_status": Comp_Rank_Status,
-                    "missing_status": Comp_Missing_Status,
-                }
-            )
-
-            horz_vars = [p_hz_data]
-            for group in horz_vars_group_cokey:
-                horz_vars.append(group.reset_index(drop=True).loc[pedon_slice_index])
-
-            dis_mat_list = []
-            for i in soil_matrix.index:
-                slice_temp = [var.loc[i] for var in horz_vars]
-                sliceT = pd.concat(slice_temp, axis=1).T
-                slice_mat = sliceT.drop(columns="compname")
-
-                if i < bedrock:
-                    sample_pedon_slice_vars = (
-                        sliceT.dropna(axis="columns").drop(columns="compname").columns.to_list()
-                    )
-                    if len(sample_pedon_slice_vars) < 2:
-                        sample_pedon_slice_vars = (
-                            sliceT[sliceT["compname"] == "sample_pedon"]
-                            .dropna(axis="columns")
-                            .drop(columns="compname")
-                            .columns.to_list()
-                        )
-                    slice_mat = sliceT[sample_pedon_slice_vars]
-
-                D = gower_distances(slice_mat)  # Equal weighting given to all soil variables
-                dis_mat_list.append(D)
-
-            # Check if any components have all NaNs at every slice
-            dis_mat_nan_check = np.ma.MaskedArray(dis_mat_list, mask=np.isnan(dis_mat_list))
-            D_check = np.ma.average(dis_mat_nan_check, axis=0)
-            rank_status = ["Not Ranked" if np.ma.is_masked(x) else "Ranked" for x in D_check[0][1:]]
-            Rank_Filter["rank_status"] = rank_status
-
-            # Calculate max dissimilarity per depth slice
-            dis_max = max(map(np.nanmax, dis_mat_list))
-
-            # Apply depth weight
-            depth_weight = np.concatenate([np.repeat(0.2, 20), np.repeat(1.0, 80)])
-            depth_weight = depth_weight[: len(soil_matrix)]
-
-            # Infill NaN data
-            for idx, dis_mat in enumerate(dis_mat_list):
-                soil_slice = soil_matrix.iloc[idx]
-                for j in range(len(dis_mat)):
-                    for k in range(len(dis_mat[j])):
-                        if np.isnan(dis_mat[j, k]):
-                            if (soil_slice[j] and not soil_slice[k]) or (
-                                not soil_slice[j] and soil_slice[k]
-                            ):
-                                dis_mat[j, k] = dis_max
-                            elif not (soil_slice[j] or soil_slice[k]):
-                                dis_mat[j, k] = 0
-
-            # Weighted average of depth-wise dissimilarity matrices
-            dis_mat_list_masked = np.ma.MaskedArray(dis_mat_list, mask=np.isnan(dis_mat_list))
-            D_sum = np.ma.average(dis_mat_list_masked, axis=0, weights=depth_weight)
-            D_sum = np.ma.filled(D_sum, fill_value=np.nan)
-            D_horz = 1 - D_sum
-
-            D_final_horz = pd.concat(
-                [
-                    compnames.reset_index(drop=True),
-                    pd.Series(D_horz[0][1:]),
-                    pd.Series(np.repeat(1.0, len(compnames))),
-                ],
-                axis=1,
-            )
-            D_final_horz.columns = ["compname", "compname_grp", "horz_score", "weight"]
-            D_final_horz = pd.merge(
-                D_final_horz,
-                mucompdata_pd[["compname", "mukey", "cokey", "distance_score", "Rank_Loc", "fss"]],
-                on="compname",
-                how="left",
-            )
-    else:
-        D_final_horz = pd.concat(
-            [
-                compnames.reset_index(drop=True),
-                pd.Series(np.repeat(0.0, len(compnames))),
-                pd.Series(np.repeat(0.0, len(compnames))),
-            ],
-            axis=1,
-        )
-        D_final_horz.columns = ["compname", "compname_grp", "horz_score", "weight"]
-        D_final_horz = pd.merge(
-            D_final_horz,
-            mucompdata_pd[["compname", "mukey", "cokey", "distance_score", "Rank_Loc", "fss"]],
-            on="compname",
-            how="left",
-        )
-
-    # ----------------------------------------------------------------------------------------
-    """
-    ##--Possibly add site disimilarity matrix in the future
-
-    if lab_Color is not None:
-        p_slope = pd.DataFrame(["sample_pedon", pSlope]).T
-        p_slope.columns = ["compname", "slope_r"]
-        soil_slope = pd.concat([p_slope, mucompdata_pd[['compname', 'slope_r']]], axis=0)
-        site_vars = pd.merge(slices_of_soil, soil_slope, on='compname', how='left')
-        compnames = site_vars[['compname']]
-        site_mat = site_vars[['slope_r', 'bottom_depth']]
-        site_mat = site_mat.set_index(compnames.compname.values)
-        D_site = gower_distances(site_mat, feature_weight=np.array([1.,1.]))
-        D_site = np.where(np.isnan(D_site), np.nanmin(D_site), D_site)
-        D_site = D_site/np.nanmax(D_site)
-        site_wt = 0.5
-        D_site = (1 - D_site)*site_wt
-        D_site_hz = np.sum([D_site, D_horz],axis=0)/(1+site_wt)
-        D_site_hz = D_site_hz/np.nanmax(D_site_hz)
-        D_final = pd.concat([compnames.reset_index(drop=True), pd.Series(D_site_hz[0][1:])], axis=1)
-    else:
-        D_final = pd.concat([compnames.reset_index(drop=True), pd.Series(D_horz[0][1:])], axis=1)
-
-    """
-    # --------------------------------------------------------------------------------------------
-
-    # Start of soil color
-
-    # Load in SRG color distribution data
-    dist1_values = [wmf1, wsf1, rmf1, rsf1, ymf1, ysf1]
-    dist2_values = [wmf2, wsf2, rmf2, rsf2, ymf2, ysf2]
-
-    data_path = current_app.config["DATA_BACKEND"]
-
-    with open(f"{data_path}/NormDist1.csv", "r") as csvfile:
-        readCSV = csv.reader(csvfile, delimiter=",")
-        for idx, row in enumerate(readCSV):
-            dist1_values[idx].extend(row)
-
-    with open(f"{data_path}/NormDist2.csv", "r") as csvfile:
-        readCSV = csv.reader(csvfile, delimiter=",")
-        for idx, row in enumerate(readCSV):
-            dist2_values[idx].extend(row)
-
-    fao74 = [
-        "Acrisols",
-        "Andosols",
-        "Arenosols",
-        "Cambisols",
-        "Chernozems",
-        "Ferralsols",
-        "Fluvisols",
-        "Gleysols",
-        "Greyzems",
-        "Histosols",
-        "Kastanozems",
-        "Luvisols",
-        "Nitosols",
-        "Phaeozems",
-        "Planosols",
-        "Podzols",
-        "Podzoluvisols",
-        "Regosols",
-        "Solonchaks",
-        "Solonetz",
-        "Vertisols",
-        "Xerosols",
-        "Yermosols",
-    ]
-
-    fao90 = [
-        "Acrisols",
-        "Alisols",
-        "Andosols",
-        "Anthrosols",
-        "Arenosols",
-        "Calcisols",
-        "Cambisols",
-        "Chernozems",
-        "Ferralsols",
-        "Fluvisols",
-        "Gleysols",
-        "Greyzems",
-        "Gypsisols",
-        "Histosols",
-        "Kastanozems",
-        "Leptosols",
-        "Lixisols",
-        "Luvisols",
-        "Nitisols",
-        "Phaeozems",
-        "Planosols",
-        "Plinthosols",
-        "Podzols",
-        "Podzoluvisols",
-        "Regosols",
-        "Solonchaks",
-        "Solonetz",
-        "Vertisols",
-    ]
-
-    # Calculate color similarity
-    if not cr_df.isnull().values.any():
-        color_sim = []
-        w_df, r_df, y_df = cr_df.iloc[0], cr_df.iloc[1], cr_df.iloc[2]
-        fao74 = [item.lower() for item in fao74]
-        fao90 = [item.lower() for item in fao90]
-
-        for compname, fss in zip(D_final_horz.compname, D_final_horz.fss):
-            soilgroup = re.sub(r"\d+$", "", " ".join(compname.split()[1:])).lower()
-
-            prob_w, prob_r, prob_y = [], [], []
-
-            if fss.lower() == "fao74":
-                fao_list, wmf, wsf, rmf, rsf, ymf, ysf = fao74, wmf1, wsf1, rmf1, rsf1, ymf1, ysf1
-            else:
-                fao_list, wmf, wsf, rmf, rsf, ymf, ysf = fao90, wmf2, wsf2, rmf2, rsf2, ymf2, ysf2
-
-            idx = fao_list.index(soilgroup) if soilgroup in fao_list else -1
-
-            for mw, sw, mr, sr, my, sy in zip(wmf, wsf, rmf, rsf, ymf, ysf):
-                prob_w.append(scipy.stats.norm(float(mw), float(sw)).pdf(float(w_df)))
-                prob_r.append(scipy.stats.norm(float(mr), float(sr)).pdf(float(r_df)))
-                prob_y.append(scipy.stats.norm(float(my), float(sy)).pdf(float(y_df)))
-
-            max_prob_w, min_prob_w = max(prob_w), min(prob_w)
-            max_prob_r, min_prob_r = max(prob_r), min(prob_r)
-            max_prob_y, min_prob_y = max(prob_y), min(prob_y)
-
-            for j in range(len(fao_list)):
-                prob_w[j] = (prob_w[j] - min_prob_w) / (max_prob_w - min_prob_w)
-                prob_r[j] = (prob_r[j] - min_prob_r) / (max_prob_r - min_prob_r)
-                prob_y[j] = (prob_y[j] - min_prob_y) / (max_prob_y - min_prob_y)
-
-            crsr = (prob_w[idx] + prob_r[idx] + prob_y[idx]) / 3.0 if idx != -1 else 1.0
-            color_sim.append(crsr)
-
-        color_sim = pd.Series(color_sim)
-
-    # Calculate Data score
-    global color_weight
-    color_weight = 0.3
-
-    if D_final_horz.horz_score.any() > 0 and cr_df.notnull().values.any():
-        D_final_horz["Score_Data"] = (D_final_horz.horz_score + (color_sim * color_weight)) / (
-            color_weight + D_final_horz.weight
-        )
-    elif D_final_horz.horz_score.all() == 0 and cr_df.notnull().values.any():
-        D_final_horz["Score_Data"] = color_sim
-        rank_status = ["Ranked with color data"] * len(mucompdata_pd)
-        missing_status = ["Color data only"] * len(mucompdata_pd)
-        Rank_Filter = pd.DataFrame(
-            {
-                "compname": mucompdata_pd.compname,
-                "rank_status": rank_status,
-                "missing_status": missing_status,
-            }
-        )
-    elif D_final_horz.horz_score.any() > 0 and cr_df.isnull().values.any():
-        D_final_horz["Score_Data"] = D_final_horz.horz_score
-    else:
-        D_final_horz["Score_Data"] = D_final_horz.horz_score
-        rank_status = ["Ranked with location data"] * len(mucompdata_pd)
-        missing_status = ["Location data only"] * len(mucompdata_pd)
-        Rank_Filter = pd.DataFrame(
-            {
-                "compname": mucompdata_pd.compname,
-                "rank_status": rank_status,
-                "missing_status": missing_status,
-            }
-        )
-
-    # Sorting and reindexing of Data-only score
-    soilIDList_data = []
-    D_final_comp_grps = [g for _, g in D_final_horz.groupby("compname_grp", sort=False)]
-
-    for comp_grps_temp in D_final_comp_grps:
-        comp_grps_temp = comp_grps_temp.sort_values("Score_Data", ascending=False).reset_index(
-            drop=True
-        )
-        SID_data = [True] + [False] * (len(comp_grps_temp) - 1)
-        comp_grps_temp["soilID_rank_data"] = SID_data
-        soilIDList_data.append(comp_grps_temp)
-
-    D_final_horz = pd.concat(soilIDList_data)
-
-    # Generate Rank_Data
-    D_final_horz = D_final_horz.sort_values(
-        ["Score_Data", "compname"], ascending=[False, True]
-    ).reset_index(drop=True)
-    rank_id = 1
-    Rank_Data = []
-
-    for is_ranked in D_final_horz["soilID_rank_data"]:
-        if is_ranked:
-            Rank_Data.append(str(rank_id))
-            rank_id += 1
-        else:
-            Rank_Data.append("Not Displayed")
-
-    D_final_horz["Rank_Data"] = Rank_Data
-
-    D_final_loc = pd.merge(
-        D_final_horz,
-        mucompdata_pd[["compname", "clay", "distance_score_norm"]],
-        on="compname",
-        how="left",
-    )
-    D_final_loc = pd.merge(D_final_loc, Rank_Filter, on="compname", how="left")
-
-    # If Score_Data returns NA, assign 0 to weight
-    D_final_loc.loc[D_final_loc.Score_Data.isnull(), "weight"] = 0
-
-    # Adjust Data Score weight by the data completeness
-    Score_Data_Loc = (D_final_loc[["Score_Data", "distance_score"]].sum(axis=1)) / (
-        D_final_loc.weight + 1
-    )
-    Score_Data_Loc = Score_Data_Loc / np.nanmax(Score_Data_Loc)
-    D_final_loc["Score_Data_Loc"] = Score_Data_Loc
-
-    # Rule-based final score adjustment
-    for i, row in D_final_loc.iterrows():
-        if cracks and row["clay"] == "Yes" and "vert" in row["compname"].lower():
-            D_final_loc.at[i, "Score_Data_Loc"] = 1.001
-        elif (
-            0 <= bedrock <= 10
-            and row["fss"].lower() == "fao74"
-            and "lithosols" in row["compname"].lower()
-        ):
-            D_final_loc.at[i, "Score_Data_Loc"] = 1.001
-        elif (
-            0 <= bedrock <= 10
-            and row["fss"].lower() == "fao90"
-            and "lithic leptosols" in row["compname"].lower()
-        ):
-            D_final_loc.at[i, "Score_Data_Loc"] = 1.001
-        elif (
-            10 < bedrock <= 30
-            and row["fss"].lower() == "fao74"
-            and ("rendzinas" in row["compname"].lower() or "rankers" in row["compname"].lower())
-        ):
-            D_final_loc.at[i, "Score_Data_Loc"] = 1.001
-        elif (
-            10 < bedrock <= 30
-            and row["fss"].lower() == "fao90"
-            and "leptosols" in row["compname"].lower()
-        ):
-            D_final_loc.at[i, "Score_Data_Loc"] = 1.001
-        elif (bedrock is None or bedrock > 50) and any(
-            term in row["compname"].lower()
-            for term in ["lithosols", "leptosols", "rendzinas", "rankers"]
-        ):
-            D_final_loc.at[i, "Score_Data_Loc"] = 0.001
-
-    D_final_loc["Score_Data_Loc"] = D_final_loc["Score_Data_Loc"] / np.nanmax(
-        D_final_loc["Score_Data_Loc"]
-    )
-    D_final_loc = D_final_loc.sort_values("Score_Data_Loc", ascending=False)
-
-    # Sorting and reindexing of final dataframe
-    soilIDList_out = []
-
-    # Group by 'compname_grp'
-    for _, comp_grps_temp in D_final_loc.groupby("compname_grp", sort=False):
-        comp_grps_temp = comp_grps_temp.sort_values("Score_Data_Loc", ascending=False).reset_index(
-            drop=True
-        )
-
-        if len(comp_grps_temp) > 1:
-            # Mark the first entry as True, all others as False
-            comp_grps_temp["soilID_rank_final"] = [True] + [False] * (len(comp_grps_temp) - 1)
-        else:
-            comp_grps_temp["soilID_rank_final"] = [True]
-
-        soilIDList_out.append(comp_grps_temp)
-
-    # Reassemble the dataframe
-    D_final_loc = pd.concat(soilIDList_out)
-
-    # Sort dataframe by score and compname
-    D_final_loc = D_final_loc.sort_values(
-        ["Score_Data_Loc", "compname"], ascending=[False, True]
-    ).reset_index(drop=True)
-
-    # Assign final rank
-    rank_id = 1
-    Rank_DataLoc = []
-    for is_rank_final in D_final_loc["soilID_rank_final"]:
-        if is_rank_final:
-            Rank_DataLoc.append(str(rank_id))
-            rank_id += 1
-        else:
-            Rank_DataLoc.append("Not Displayed")
-
-    D_final_loc["Rank_Data_Loc"] = Rank_DataLoc
-    D_final_loc = D_final_loc.sort_values(
-        ["soilID_rank_final", "Score_Data_Loc"], ascending=[False, False]
-    ).reset_index(drop=True)
-    D_final_loc[
-        ["Score_Data", "horz_score", "Score_Data_Loc", "distance_score", "distance_score_norm"]
-    ] = D_final_loc[
-        ["Score_Data", "horz_score", "Score_Data_Loc", "distance_score", "distance_score_norm"]
-    ].fillna(
-        0.0
-    )
-
-    # Final scaling of similarity score by 'data_completeness'
-    D_final_loc.Score_Data_Loc = D_final_loc.Score_Data_Loc * (
-        0.25 if data_completeness < 25 else float(data_completeness) / 100
-    )
-
-    # Constructing rank list
-    rank_list = []
-
-    is_location_only = D_final_loc.missing_status.all() == "Location data only"
-    for idx, row in D_final_loc.iterrows():
-        if is_location_only:
-            rank_entry = {
-                "name": row.compname.capitalize(),
-                "component": row.compname_grp.capitalize(),
-                "mapunitID": row.mukey,
-                "componentID": row.cokey,
-                "score_data_loc": "",
-                "rank_data_loc": "",
-                "score_data": "",
-                "rank_data": "",
-                "score_loc": round(row.distance_score_norm, 3),
-                "rank_loc": row.Rank_Loc,
-                "componentData": row.missing_status,
-            }
-        else:
-            rank_entry = {
-                "name": row.compname.capitalize(),
-                "component": row.compname_grp.capitalize(),
-                "mapunitID": row.mukey,
-                "componentID": row.cokey,
-                "score_data_loc": round(row.Score_Data_Loc, 3),
-                "rank_data_loc": row.Rank_Data_Loc,
-                "score_data": round(row.Score_Data, 3),
-                "rank_data": row.Rank_Data,
-                "score_loc": round(row.distance_score_norm, 3),
-                "rank_loc": row.Rank_Loc,
-                "componentData": row.missing_status,
-            }
-        rank_list.append(rank_entry)
-
-    # Setting up the return data structure
-    model_version = 3
-    metadata = {
-        "location": "global",
-        "model": "v3",
-        "dataCompleteness": {"score": data_completeness, "text": text_completeness},
-    }
-
-    result = {"metadata": metadata, "soilRank": rank_list}
-
-    # Save data if record_id is provided
-    if record_id:
-        save_rank_output(record_id, model_version, json.dumps(result))
-
-    return result
-
-
-#####################################################################################################
-#                                   getSoilLocationBasedUS                                          #
-#####################################################################################################
+############################################################################################
+#                                   getSoilLocationBasedUS                                 #
+############################################################################################
 def getSoilLocationBasedUS(lon, lat, plot_id):
     # Load in LAB to Munsell conversion look-up table
     color_ref = pd.read_csv("%s/LandPKS_munsell_rgb_lab.csv" % current_app.config["DATA_BACKEND"])
+    LAB_ref = color_ref[["L", "A", "B"]]
+    # rgb_ref = color_ref[["r", "g", "b"]]
+    munsell_ref = color_ref[["hue", "value", "chroma"]]
 
     # Load in SSURGO data from SoilWeb
-    # soilweb_url = f"https://casoilresource.lawr.ucdavis.edu/api/landPKS.php?q=spn&lon={lon}&lat={lat}&r=1000" # current production API
-    soilweb_url = f"https://soilmap2-1.lawr.ucdavis.edu/dylan/soilweb/api/landPKS.php?q=spn&lon={lon}&lat={lat}&r=1000"  # testing API
+    # current production API
+    # soilweb_url = f"""https://casoilresource.lawr.ucdavis.edu/api/landPKS.php?q=spn&lon={lon}&lat={lat}&r=1000"""
+
+    # testing API
+    soilweb_url = f"""https://soilmap2-1.lawr.ucdavis.edu/dylan/soilweb/api/landPKS.php?q=spn&lon={lon}&lat={lat}&r=1000"""
     try:
         response = requests.get(soilweb_url, timeout=8)
         out = response.json()
     except requests.RequestException as e:
         print(f"Error fetching data from SoilWeb: {e}")
-        out = {"ESD": False, "OSD_morph": False, "OSD_narrative": False, "hz": False, "spn": False}
+        out = {
+            "ESD": False,
+            "OSD_morph": False,
+            "OSD_narrative": False,
+            "hz": False,
+            "spn": False,
+        }
 
     OSD_compkind = ["Series", "Variant", "Family", "Taxadjunct"]
     # Check if point is in a NOTCOM area, and if so then infill with STATSGO from NRCS SDA
@@ -1302,7 +173,11 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
         # Create a DataFrame for distances and intersections
         mu_id_dist = pd.DataFrame(
-            {"MUKEY": mu_geo["MUKEY"].values, "distance": distances, "pt_intersect": intersects}
+            {
+                "MUKEY": mu_geo["MUKEY"].values,
+                "distance": distances,
+                "pt_intersect": intersects,
+            }
         )
 
         # Update distance to 0 for intersecting mapunits
@@ -1314,10 +189,13 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
         # Build the mucompdata query
         mukey_list = mukey_dist_final["MUKEY"].tolist()
-        mucompdataQry = f"""SELECT component.mukey, component.cokey, component.compname, component.comppct_r, 
-        component.compkind, component.majcompflag, component.slope_r, component.elev_r, component.nirrcapcl, 
-        component.nirrcapscl, component.nirrcapunit, component.irrcapcl, component.irrcapscl, component.irrcapunit, 
-        component.taxorder, component.taxsubgrp FROM component WHERE mukey IN ({','.join(map(str, mukey_list))})"""
+        mucompdataQry = f"""SELECT component.mukey, component.cokey, component.compname,
+        component.comppct_r, component.compkind, component.majcompflag,
+        component.slope_r, component.elev_r, component.nirrcapcl, component.nirrcapscl,
+        component.nirrcapunit, component.irrcapcl, component.irrcapscl,
+        component.irrcapunit, component.taxorder, component.taxsubgrp
+        FROM component WHERE mukey IN ({','.join(map(str, mukey_list))})"""
+
         mucompdata_out = sda_return(propQry=mucompdataQry)
 
         # Process the mucompdata results
@@ -1408,21 +286,21 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     # Process and fill missing or zero values in the 'comppct_r' column.
     mucompdata_pd = fill_missing_comppct_r(mucompdata_pd)
 
-    # --------------------------------------------------------------------------------
+    # --------------------------------------------------------------------
     # Location based calculation
     # -----------------------------
-
-    # ---------------------------------------------------------------------------------
-    ###################################################################################
-    # Individual probability
-    # Based on Fan et al 2018 EQ 1, the conditional probability for each component is 
-    # calculated by taking the sum of all occurances of a component in the home and 
-    # adjacent mapunits and dividing this by the sum of all map units and components. 
-    # We have modified this approach so that each instance of a component occurance is 
-    # evaluated separately and assinged a weight and the max distance score for each 
-    # component group is assignedto all component instances.
-    ###################################################################################
-
+    """
+    # --------------------------------------------------------------------
+    ############################################################################
+    Individual probability
+    Based on Fan et al 2018 EQ 1, the conditional probability for each component
+    is calculated by taking the sum of all occurances of a component in the
+    home and adjacent mapunits and dividing this by the sum of all map units
+    and components. We have modified this approach so that each instance of a
+    component occurance is evaluated separately and assinged a weight and the
+    max distance score for each component group is assigned to all component instances.
+    ############################################################################
+    """
     # Process distance scores and perform group-wise aggregations.
     mucompdata_pd = process_distance_scores(mucompdata_pd, ExpCoeff)
 
@@ -1437,9 +315,9 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     # Get component key list
     comp_key = mucompdata_pd["cokey"].tolist()
 
-    # --------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
     # Extracts horizon data
-    # --------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
     if data_source == "SSURGO":
         # Convert JSON data to DataFrame
         muhorzdata_pd = pd.json_normalize(out["hz"])[
@@ -1449,9 +327,24 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                 "hzdepb_r",
                 "chkey",
                 "hzname",
+                "sandtotal_l",
                 "sandtotal_r",
+                "sandtotal_h",
+                "silttotal_l",
                 "silttotal_r",
+                "silttotal_h",
+                "claytotal_l",
                 "claytotal_r",
+                "claytotal_h",
+                "dbovendry_l",
+                "dbovendry_r",
+                "dbovendry_h",
+                "wthirdbar_l",
+                "wthirdbar_r",
+                "wthirdbar_h",
+                "wfifteenbar_l",
+                "wfifteenbar_r",
+                "wfifteenbar_h",
                 "total_frag_volume",
                 "cec7_r",
                 "ecec_r",
@@ -1460,7 +353,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                 "lep_r",
             ]
         ]
-
         # Convert specific columns to appropriate data types
         numeric_columns = [
             "total_frag_volume",
@@ -1509,19 +401,21 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                 # Process and fill missing or zero values in the 'comppct_r' column.
                 mucompdata_pd = fill_missing_comppct_r(mucompdata_pd)
 
-                # --------------------------------------------------------------------------------
-                # Location based calculation - Run on STATSGO b/c SSURGO horizon data was NULL
+                # --------------------------------------------------------------------
+                # Location based calculation
                 # -----------------------------
-                ###################################################################################
-                # Individual probability
-                # Based on Fan et al 2018 EQ 1, the conditional probability for each component is 
-                # calculated by taking the sum of all occurances of a component in the home and 
-                # adjacent mapunits and dividing this by the sum of all map units and components. 
-                # We have modified this approach so that each instance of a component occurance is 
-                # evaluated separately and assinged a weight and the max distance score for each 
-                # component group is assignedto all component instances.
-                ###################################################################################
-
+                """
+                # --------------------------------------------------------------------
+                ############################################################################
+                Individual probability
+                Based on Fan et al 2018 EQ 1, the conditional probability for each component
+                is calculated by taking the sum of all occurances of a component in the
+                home and adjacent mapunits and dividing this by the sum of all map units
+                and components. We have modified this approach so that each instance of a
+                component occurance is evaluated separately and assinged a weight and the
+                max distance score for each component group is assigned to all component instances.
+                ############################################################################
+                """
                 # Process distance scores and perform group-wise aggregations.
                 mucompdata_pd = process_distance_scores(mucompdata_pd, ExpCoeff)
 
@@ -1535,7 +429,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                     "cokey"
                 ].tolist()
 
-                # ------------------------------------------------------------------------------------
+                # -----------------------------------------------------------------------------------------------------------
                 # STATSGO Horizon Data Query
                 muhorzdata_pd = extract_muhorzdata_STATSGO(mucompdata_pd)
 
@@ -1546,7 +440,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     # Merge muhorzdata_pd with selected columns from mucompdata_pd
     muhorzdata_pd = pd.merge(
         muhorzdata_pd,
-        mucompdata_pd[["cokey", "comppct_r", "compname", "slope_r"]],
+        mucompdata_pd[["cokey", "comppct_r", "compname", "distance_score", "slope_r"]],
         on="cokey",
         how="left",
     )
@@ -1554,7 +448,8 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     # Replace "NULL" strings with numpy NaN
     muhorzdata_pd.replace("NULL", np.nan, inplace=True)
 
-    # Filter out components with missing horizon depth data that aren't either a Series, Variant, or Family
+    # Filter out components with missing horizon depth data that aren't either
+    # a Series, Variant, or Family
     filter_condition = muhorzdata_pd["cokey"].isin(cokey_series) | (
         pd.notnull(muhorzdata_pd["hzdept_r"]) & pd.notnull(muhorzdata_pd["hzdepb_r"])
     )
@@ -1627,6 +522,10 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     muhorzdata_pd = muhorzdata_pd.merge(
         mucompdata_pd[["cokey", "compname"]], on="cokey", how="left"
     )
+
+    # Remove bedrock by filtering out 'R|r' in hzname
+    muhorzdata_pd = muhorzdata_pd[~muhorzdata_pd["hzname"].str.contains("R", case=False, na=False)]
+
     # Group data by cokey
     muhorzdata_group_cokey = [group for _, group in muhorzdata_pd.groupby("cokey", sort=False)]
 
@@ -1752,7 +651,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
         cokeyB = getProfile_cokey_temp2["cokey"].iloc[0]
         compnameB = getProfile_cokey_temp2["compname"].iloc[0]
-        comppctB = getProfile_cokey_temp2["comppct"].iloc[0]
         c_bottom_depths_temp = pd.DataFrame([cokeyB, compnameB, int(c_very_bottom)]).T
         c_bottom_depths_temp.columns = ["cokey", "compname", "c_very_bottom"]
 
@@ -1767,36 +665,16 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         clay_texture_temp.columns = ["compname", "clay"]
         clay_texture.append(clay_texture_temp)
 
-        # calculate LPKS depth aggregated data based on original depth sliced horizons
-        snd_d, hz_depb = agg_data_layer(
-            data=sand_pct_intpl["c_sandpct_intpl"],
-            bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0],
-            depth=True,
-        )
-        cly_d = agg_data_layer(
-            data=clay_pct_intpl["c_claypct_intpl"],
-            bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0],
-        )
-        txt_d = []
-        for l in range(len(snd_d)):
-            text_T = getTexture(
-                row=None, sand=snd_d[l], silt=(100 - (snd_d[l] + cly_d[l])), clay=cly_d[l]
-            )
-            txt_d.append(text_T)
-        txt_d = pd.Series(txt_d)
-        txt_d.index = snd_d.index
-        rf_d = agg_data_layer(
-            data=cf_pct_intpl["c_cfpct_intpl"], bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0]
-        )
-        cec_d = agg_data_layer(
-            data=cec_intpl["c_cec_intpl"], bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0]
-        )
-        ph_d = agg_data_layer(
-            data=ph_intpl["c_ph_intpl"], bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0]
-        )
-        ec_d = agg_data_layer(
-            data=ec_intpl["c_ec_intpl"], bottom=c_bottom_depths_temp["c_very_bottom"].iloc[0]
-        )
+        # extract horizon data
+        hz_depb = group_sorted["hzdepb_r"]
+        snd_d = group_sorted["sandtotal_r"]
+        cly_d = group_sorted["claytotal_r"]
+        txt_d = group_sorted["texture"]
+        rf_d = group_sorted["total_frag_volume"]
+        cec_d = group_sorted["CEC"]
+        ec_d = group_sorted["EC"]
+        ph_d = group_sorted["pH"]
+
         hz_depb = hz_depb.fillna("")
         snd_d = snd_d.fillna("")
         cly_d = cly_d.fillna("")
@@ -1836,23 +714,244 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     comp_key = mucompdata_pd["cokey"].unique().tolist()
     cokey_Index = {key: index for index, key in enumerate(comp_key)}
 
-    # -------------------------------------------------------------------------------------------------------------
+    # ------------------------------------------------------------------------
+
+    """
+    Simulation modeling Steps:
+      Step 1. Calculate a local soil property correlation matrix uisng the
+              representative values from SSURGO.
+
+      Step 2. Steps performed on each row:
+        a. Simulate sand/silt/clay percentages using the 'simulate_correlated_triangular'
+           function, using the global correlation matrix and the local l,r,h values for
+           each particle fraction. Format as a composition using 'acomp'
+        b. Perform the isometric log-ratio transformation.
+        c. Extract l,r,h values (min, median, max for ilr1 and ilr2) and format into a
+           params object for simiulation.
+        d. Simulate all properties and then permorm inverse transform on ilr1 and ilr2
+           to obtain sand, silt, and clay values.
+        e. Append simulated values to dataframe
+
+      Step 3. Run Rosetta and other Van Genuchten equations to calcuate AWS in top 50
+              cm using simulated dataframe.
+    """
+
+    # Step 1. Calculate a local soil property correlation matrix
+
+    # Subset data based on required soil inputs
+    sim_columns = [
+        "compname_grp",
+        "distance_score",
+        "hzdept_r",
+        "hzdepb_r",
+        "sandtotal_l",
+        "sandtotal_r",
+        "sandtotal_h",
+        "silttotal_l",
+        "silttotal_r",
+        "silttotal_h",
+        "claytotal_l",
+        "claytotal_r",
+        "claytotal_h",
+        "dbovendry_l",
+        "dbovendry_r",
+        "dbovendry_h",
+        "wthirdbar_l",
+        "wthirdbar_r",
+        "wthirdbar_h",
+        "wfifteenbar_l",
+        "wfifteenbar_r",
+        "wfifteenbar_h",
+    ]
+
+    sim_data_columns = [
+        "hzdept_r",
+        "hzdepb_r",
+        "sandtotal_l",
+        "sandtotal_r",
+        "sandtotal_h",
+        "silttotal_l",
+        "silttotal_r",
+        "silttotal_h",
+        "claytotal_l",
+        "claytotal_r",
+        "claytotal_h",
+        "dbovendry_l",
+        "dbovendry_r",
+        "dbovendry_h",
+        "wthirdbar_l",
+        "wthirdbar_r",
+        "wthirdbar_h",
+        "wfifteenbar_l",
+        "wfifteenbar_r",
+        "wfifteenbar_h",
+    ]
+    sim = muhorzdata_pd[sim_columns]
+
+    # infill missing data
+    sim = infill_soil_data(sim)
+
+    agg_data = []
+
+    # Group data by compname_grp
+    sim_group_compname = [group for _, group in sim.groupby("compname_grp", sort=False)]
+    for group in sim_group_compname:
+        # aggregate data into 0-20 and 20-50 or bottom depth
+        group_ag = slice_and_aggregate_soil_data(group[sim_data_columns])
+        group_ag["compname_grp"] = group["compname_grp"].unique()[0]
+        group_ag["distance_score"] = group["distance_score"].unique()[0]
+        group_ag = group_ag.drop('Depth', axis=1)
+        agg_data.append(group_ag)
+
+    # Concatenate the results for each column into a single dataframe
+    agg_data_df = pd.concat(agg_data, axis=0, ignore_index=True).dropna().reset_index(drop=True)
+
+    # return(jsonify(agg_data_df.to_dict()))
+    # return(json.dumps(agg_data_df[["sandtotal_r", "silttotal_r", "claytotal_r"]].values.tolist()))
+
+    # Extract columns with names ending in '_r'
+    agg_data_r = agg_data_df[[col for col in agg_data_df.columns if col.endswith("_r")]]
+
+    # Compute the local correlation matrix (Spearman correlation matrix)
+    rep_columns = agg_data_r.drop(columns=["sandtotal_r", "silttotal_r", "claytotal_r"])
+    # correlation_matrix, _ = spearmanr(selected_columns, axis=0)
+
+    ilr_site_txt = ilr(agg_data_df[["sandtotal_r", "silttotal_r", "claytotal_r"]].values)
+    # return(json.dumps(ilr_site_txt[:, 0].tolist()))
+    rep_columns["ilr1"] = pd.Series(ilr_site_txt[:, 0])
+    rep_columns["ilr2"] = pd.Series(ilr_site_txt[:, 1])
+
+    correlation_matrix_data = rep_columns[
+        [
+            "ilr1",
+            "ilr2",
+            "dbovendry_r",
+            "wthirdbar_r",
+            "wfifteenbar_r",
+        ]
+    ]
+    local_correlation_matrix, _ = spearmanr(correlation_matrix_data, axis=0)
+    """
+    Step 2. Simulate data for each row, with the number of simulations equal
+            to the (distance_score*100)*10
+    """
+    # Global soil texture correlation matrix (used for initial simulation)
+    texture_correlation_matrix = np.array(
+        [
+            [1.0000000, -0.76231798, -0.67370589],
+            [-0.7623180, 1.00000000, 0.03617498],
+            [-0.6737059, 0.03617498, 1.00000000],
+        ]
+    )
+
+    sim_data_out = []
+
+    for _, row in agg_data_df.iterrows():
+        # 2a. Simulate sand/silt/clay percentages
+        # 1. Extract and format data params
+        sand_params = [row["sandtotal_l"], row["sandtotal_r"], row["sandtotal_h"]]
+        silt_params = [row["silttotal_l"], row["silttotal_r"], row["silttotal_h"]]
+        clay_params = [row["claytotal_l"], row["claytotal_r"], row["claytotal_h"]]
+
+        params_txt = [sand_params, silt_params, clay_params]
+
+        # 2. Perform processing steps on data
+        #    Convert simulated data using the acomp function and then compute the isometric
+        #    log-ratio transformation.
+        simulated_txt = acomp(
+            simulate_correlated_triangular(
+                (int(row["distance_score"] * 1000)), params_txt, texture_correlation_matrix
+            )
+        )
+        simulated_txt_ilr = ilr(simulated_txt)
+
+        # Extract min, median, and max for the first two ilr transformed columns.
+        ilr1_values = simulated_txt_ilr[:, 0]
+        ilr2_values = simulated_txt_ilr[:, 1]
+
+        ilr1_l, ilr1_r, ilr1_h = (
+            ilr1_values.min(),
+            np.median(ilr1_values),
+            ilr1_values.max(),
+        )
+        ilr2_l, ilr2_r, ilr2_h = (
+            ilr2_values.min(),
+            np.median(ilr2_values),
+            ilr2_values.max(),
+        )
+
+        # Create the list of parameters.
+        params = [
+            [ilr1_l, ilr1_r, ilr1_h],
+            [ilr2_l, ilr2_r, ilr2_h],
+            [row["dbovendry_l"], row["dbovendry_r"], row["dbovendry_h"]],
+            [row["wthirdbar_l"], row["wthirdbar_r"], row["wthirdbar_h"]],
+            [row["wfifteenbar_l"], row["wfifteenbar_r"], row["wfifteenbar_h"]],
+        ]
+
+        sim_data = simulate_correlated_triangular(
+            int(row["distance_score"] * 1000), params, local_correlation_matrix
+        )
+        sim_data = pd.DataFrame(
+            sim_data,
+            columns=[
+                "ilr1",
+                "ilr2",
+                "bulk_density_third_bar",
+                "water_retention_third_bar",
+                "water_retention_15_bar",
+            ],
+        )
+        sim_data['water_retention_third_bar'] = sim_data['water_retention_third_bar'].div(100)
+        sim_data['water_retention_15_bar'] = sim_data['water_retention_15_bar'].div(100)
+        sim_txt = ilr_inv(sim_data[["ilr1", "ilr2"]])
+        sim_txt = pd.DataFrame(sim_txt, columns=["sand_total", "silt_total", "clay_total"])
+        sim_txt = sim_txt.multiply(100)
+        multi_sim = pd.concat([sim_data.drop(columns=["ilr1", "ilr2"]), sim_txt], axis=1)
+        multi_sim['compname'] = row['compname']
+        multi_sim['hzdept_r'] = row['hzdept_r']
+        multi_sim['hzdepb_r'] = row['hzdepb_r']
+        sim_data_out.append(multi_sim)
+    sim_data_df = pd.concat(sim_data_out, axis=0, ignore_index=True)
+
+    # Convert NaN values to None
+    sim_data_df = sim_data_df.where(pd.notna(sim_data_df), None)
+
+    # Convert the DataFrame to the desired format
+    # rossetta_input_data = sim_data_df.values.tolist()
+    variables = [
+        "sand_total",
+        "silt_total",
+        "clay_total",
+        "bulk_density_third_bar",
+        "water_retention_third_bar",
+        "water_retention_15_bar",
+    ]
+    rosetta_data = process_data_with_rosetta(sim_data_df, vars=variables, v="3")
+
+    # Create layerID
+    sim_data_df['layerID'] = sim_data_df['compname_grp'] + "_" + sim_data_df['hzdept_r'].astype(str)
+    # return jsonify(sim_data_df.to_dict())
+
+    # ----------------------------------------------------------------------------
     # This extracts OSD color, texture, and CF data
     if data_source == "STATSGO":
-        # If the condition is met, we perform the series of operations, otherwise, we set OSDhorzdata_pd to None
+        # If the condition is met, we perform the series of operations, otherwise,
+        # we set OSDhorzdata_pd to None
         if mucompdata_pd["compkind"].isin(OSD_compkind).any():
             try:
                 # Generate series names
                 series_name = [
-                    urllib.parse.quote(re.sub("[0-9]+", "", compname))
+                    re.sub("[0-9]+", "", compname)
                     for compname in mucompdata_pd["compname"]
                     if compname in OSD_compkind
                 ]
-                series_join = ",".join(series_name)
+
+                params = {"q": "site_hz", "s": series_name}
 
                 # Fetch data from URL using requests
-                series_url = f"https://casoilresource.lawr.ucdavis.edu/api/soil-series.php?q=site_hz&s={series_join}"
-                response = requests.get(series_url, timeout=3)
+                series_url = "https://casoilresource.lawr.ucdavis.edu/api/soil-series.php"
+                response = requests.get(series_url, params=params, timeout=3)
                 seriesDict = response.json()
 
                 # Normalize the data and perform data manipulations
@@ -1875,7 +974,8 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                     .str.replace(r"(fine|medium|coarse) ", "")
                 )
                 OSDhorzdata_pd.loc[
-                    OSDhorzdata_pd["matrix_dry_color_hue"] == "N", "matrix_dry_color_value"
+                    OSDhorzdata_pd["matrix_dry_color_hue"] == "N",
+                    "matrix_dry_color_value",
                 ] = 1
 
                 # Color conversion
@@ -1925,6 +1025,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                 OSDhorzdata_pd["sandtotal_r"] = OSDhorzdata_pd["texture_class"].apply(getSand)
             except Exception as err:
                 OSDhorzdata_pd = None
+                print("An error occurred:", err)
         else:
             OSDhorzdata_pd = None
 
@@ -1992,7 +1093,9 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     base_columns = ["mukey", "cokey", "compname"]
     base_df = mucompdata_pd[base_columns]
     nan_columns = pd.DataFrame(
-        np.nan, index=np.arange(len(mucompdata_pd.cokey)), columns=["series", "brief_narrative"]
+        np.nan,
+        index=np.arange(len(mucompdata_pd.cokey)),
+        columns=["series", "brief_narrative"],
     )
 
     # Conditions for 'STATSGO' or absence of 'OSD_narrative' in 'SSURGO'
@@ -2025,6 +1128,11 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             inplace=True,
         )
 
+        # Remove bedrock by filtering out 'R|r' in hzname
+        OSDhorzdata_pd = OSDhorzdata_pd[
+            ~OSDhorzdata_pd["hzname"].str.contains("R", case=False, na=False)
+        ]
+
         if mucompdata_pd["compkind"].isin(OSD_compkind).any():
             # Group data by cokey
             OSDhorzdata_group_cokey = [
@@ -2035,7 +1143,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             lab_lyrs = []
             munsell_lyrs = []
             lab_intpl_lyrs = []
-            osd_hz_lyrs = []
 
             # Define helper functions outside the loop
             def create_new_layer(row, top, bottom):
@@ -2052,11 +1159,18 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                     "matrix_dry_color_chroma",
                 ]:
                     new_row[col] = None
-                for col in ["r", "g", "b", "total_frag_volume", "claytotal_r", "sandtotal_r"]:
+                for col in [
+                    "r",
+                    "g",
+                    "b",
+                    "total_frag_volume",
+                    "claytotal_r",
+                    "sandtotal_r",
+                ]:
                     new_row[col] = np.nan
                 return new_row
 
-            for group in OSDhorzdata_group_cokey:
+            for index, group in enumerate(OSDhorzdata_group_cokey):
                 group_sorted = group.sort_values(by="top").drop_duplicates().reset_index(drop=True)
 
                 # Remove invalid horizons where top depth is greater than bottom depth
@@ -2109,12 +1223,12 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                     c_bottom_depths_group = c_bottom_depths[
                         c_bottom_depths["cokey"].isin(group_sorted["cokey"])
                     ]
-                    # return(jsonify(c_bottom_depths_group.to_dict(orient='records')))
+                    muhorzdata_pd_group = muhorzdata_pd[
+                        muhorzdata_pd["cokey"].isin(group_sorted["cokey"])
+                    ]
+
                     # Check if OSD depth adjustment is needed
-                    if (
-                        OSD_very_bottom < c_bottom_depths_group["c_very_bottom"].iloc[0]
-                        and OSD_very_bottom < 120
-                    ):
+                    if OSD_very_bottom < c_bottom_depths_group["c_very_bottom"].iloc[0]:
                         OSD_depth_add = True
                         depth_difference = (
                             c_bottom_depths_group["c_very_bottom"].iloc[0] - OSD_very_bottom
@@ -2122,6 +1236,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                         lab_values = [
                             lab_intpl.loc[OSD_very_bottom - 1].values.tolist()
                         ] * depth_difference
+
                         pd_add = pd.DataFrame(lab_values, columns=["l", "a", "b"])
 
                         # Adjust LAB values for the OSD depth
@@ -2130,7 +1245,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                         ).reset_index(drop=True)
                         OSD_very_bottom = c_bottom_depths_group["c_very_bottom"].iloc[0]
 
-                    elif 0 < c_bottom_depths_group["c_very_bottom"].iloc[0] < 120 < OSD_very_bottom:
+                    elif 0 < c_bottom_depths_group["c_very_bottom"].iloc[0] < OSD_very_bottom:
                         OSD_depth_remove = True
 
                         # Adjust LAB values for the component depth
@@ -2146,24 +1261,28 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                     # If all values in lab_intpl are null, append default values to lists
                     if lab_intpl.isnull().values.all():
                         lab_lyrs.append(["", "", ""])
-                        osd_hz_lyrs.append("")
                         munsell_lyrs.append("")
                     else:
                         # Aggregate data for each color dimension
-                        l_d, osd_hz_d = agg_data_layer(
-                            data=lab_intpl["l"], bottom=OSD_very_bottom, sd=4, depth=True
-                        )
-                        a_d = agg_data_layer(
-                            data=lab_intpl["a"], bottom=OSD_very_bottom, sd=4
+                        l_d = aggregate_data(
+                            data=lab_intpl["l"],
+                            bottom_depths=muhorzdata_pd_group["hzdepb_r"].tolist(),
+                            sd=2,
                         ).fillna("")
-                        b_d = agg_data_layer(
-                            data=lab_intpl["b"], bottom=OSD_very_bottom, sd=4
+                        a_d = aggregate_data(
+                            data=lab_intpl["a"],
+                            bottom_depths=muhorzdata_pd_group["hzdepb_r"].tolist(),
+                            sd=2,
+                        ).fillna("")
+                        b_d = aggregate_data(
+                            data=lab_intpl["b"],
+                            bottom_depths=muhorzdata_pd_group["hzdepb_r"].tolist(),
+                            sd=2,
                         ).fillna("")
 
                         # Convert LAB values to a list of triplets
-                        lab_parse = [[l, a, b] for l, a, b in zip(l_d, a_d, b_d)]
+                        lab_parse = [[L, A, B] for L, A, B in zip(l_d, a_d, b_d)]
                         lab_lyrs.append(dict(zip(l_d.index, lab_parse)))
-                        osd_hz_lyrs.append(dict(zip(osd_hz_d.index, osd_hz_d)))
 
                         # Convert LAB triplets to Munsell values
                         munsell_values = [
@@ -2173,19 +1292,25 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                             for lab in lab_parse
                         ]
                         munsell_lyrs.append(dict(zip(l_d.index, munsell_values)))
-
+                    # return(jsonify(OSD_text_int  = str(OSD_text_int[index] )))
+                    # return(jsonify(group_sorted.to_dict(orient='records')))
                     # Extract OSD Texture and Rock Fragment Data
-                    if OSD_text_int[i] == "Yes" or OSD_rfv_int[i] == "Yes":
+                    if OSD_text_int[index] == "Yes" or OSD_rfv_int[index] == "Yes":
                         group_sorted[["hzdept_r", "hzdepb_r", "texture"]] = group_sorted[
                             ["top", "bottom", "texture_class"]
                         ]
-
                         OSD_very_bottom_int, OSD_clay_intpl = getProfile(
                             group_sorted, "claytotal_r", c_bot=True
                         )
-                        OSD_clay_intpl.columns = ["c_claypct_intpl", "c_claypct_intpl_grp"]
+                        OSD_clay_intpl.columns = [
+                            "c_claypct_intpl",
+                            "c_claypct_intpl_grp",
+                        ]
                         OSD_sand_intpl = getProfile(group_sorted, "sandtotal_r")
-                        OSD_sand_intpl.columns = ["c_sandpct_intpl", "c_sandpct_intpl_grp"]
+                        OSD_sand_intpl.columns = [
+                            "c_sandpct_intpl",
+                            "c_sandpct_intpl_grp",
+                        ]
                         OSD_rfv_intpl = getProfile(group_sorted, "total_frag_volume")
                         OSD_rfv_intpl.columns = ["c_cfpct_intpl", "c_cfpct_intpl_grp"]
 
@@ -2226,12 +1351,13 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                             OSD_very_bottom,
                         )
 
-                        # If OSD bottom depth is greater than component depth and component depth is <120cm
+                        # If OSD bottom depth is greater than component depth and component depth
+                        # is <120cm
                         if OSD_depth_remove:
                             # Remove data based on c_bottom_depths
-                            OSD_sand_intpl = OSD_sand_intpl.loc[: c_bottom_depths.iloc[i, 2]]
-                            OSD_clay_intpl = OSD_clay_intpl.loc[: c_bottom_depths.iloc[i, 2]]
-                            OSD_rfv_intpl = OSD_rfv_intpl.loc[: c_bottom_depths.iloc[i, 2]]
+                            OSD_sand_intpl = OSD_sand_intpl.loc[: c_bottom_depths.iloc[index, 2]]
+                            OSD_clay_intpl = OSD_clay_intpl.loc[: c_bottom_depths.iloc[index, 2]]
+                            OSD_rfv_intpl = OSD_rfv_intpl.loc[: c_bottom_depths.iloc[index, 2]]
 
                         # Create the compname and cokey dataframes
                         compname_df = pd.DataFrame(
@@ -2259,7 +1385,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                         ]
 
                         # Update getProfile_mod based on conditions
-                        getProfile_mod = getProfile_cokey[i]
+                        getProfile_mod = getProfile_cokey[index]
                         compname_check = (
                             getProfile_mod["compname"]
                             .isin(group_sorted2[["compname"]].iloc[0])
@@ -2268,7 +1394,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
                         if (
                             compname_check
-                            and OSD_text_int[i] == "Yes"
+                            and OSD_text_int[index] == "Yes"
                             and not group_sorted2["c_sandpct_intpl"].isnull().all()
                         ):
                             getProfile_mod["sandpct_intpl"] = group_sorted2["c_sandpct_intpl"]
@@ -2276,16 +1402,18 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
                         if (
                             compname_check
-                            and OSD_rfv_int[i] == "Yes"
+                            and OSD_rfv_int[index] == "Yes"
                             and not group_sorted2["c_cfpct_intpl"].isnull().all()
                         ):
                             getProfile_mod["rfv_intpl"] = group_sorted2["c_cfpct_intpl"]
 
-                        getProfile_cokey[i] = getProfile_mod
+                        getProfile_cokey[index] = getProfile_mod
 
                         # Aggregate sand data
                         snd_d_osd, hz_depb_osd = agg_data_layer(
-                            data=OSD_sand_intpl.iloc[:, 0], bottom=OSD_very_bottom, depth=True
+                            data=OSD_sand_intpl.iloc[:, 0],
+                            bottom=OSD_very_bottom,
+                            depth=True,
                         )
 
                         # Aggregate clay data
@@ -2313,43 +1441,42 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
                         hz_depb_osd.fillna("", inplace=True)
 
                         # Store aggregated data in dictionaries based on conditions
-                        if OSD_text_int[i] == "Yes":
-                            snd_lyrs[i] = snd_d_osd.to_dict()
-                            cly_lyrs[i] = cly_d_osd.to_dict()
-                            txt_lyrs[i] = txt_d_osd.to_dict()
+                        if OSD_text_int[index] == "Yes":
+                            snd_lyrs[index] = snd_d_osd.to_dict()
+                            cly_lyrs[index] = cly_d_osd.to_dict()
+                            txt_lyrs[index] = txt_d_osd.to_dict()
 
-                        if OSD_rfv_int[i] == "Yes":
-                            rf_lyrs[i] = rf_d_osd.to_dict()
+                        if OSD_rfv_int[index] == "Yes":
+                            rf_lyrs[index] = rf_d_osd.to_dict()
 
                         # Update horizon layers if bottom depth is zero
-                        if c_bottom_depths.iloc[i, 2] == 0:
-                            hz_lyrs[i] = hz_depb_osd.to_dict()
-                            mucompdata_pd.loc[i, "c_very_bottom"] = OSD_very_bottom
+                        if c_bottom_depths.iloc[index, 2] == 0:
+                            hz_lyrs[index] = hz_depb_osd.to_dict()
+                            mucompdata_pd.loc[index, "c_very_bottom"] = OSD_very_bottom
 
                         # Update cec, ph, and ec layers if they contain only a single empty string
                         for lyr in [cec_lyrs, ph_lyrs, ec_lyrs]:
-                            if len(lyr[i]) == 1 and lyr[i][0] == "":
-                                lyr[i] = dict(zip(hz_depb_osd.index, [""] * len(hz_depb_osd)))
+                            if len(lyr[index]) == 1 and lyr[index][0] == "":
+                                lyr[index] = dict(zip(hz_depb_osd.index, [""] * len(hz_depb_osd)))
 
                 else:
-                    OSDhorzdata_group_cokey[i] = group_sorted
+                    OSDhorzdata_group_cokey[index] = group_sorted
 
                     # Create an empty dataframe with NaNs for lab_intpl
                     lab_intpl = pd.DataFrame(
-                        np.nan, index=np.arange(c_bottom_depths.iloc[i, 2]), columns=["l", "a", "b"]
+                        np.nan,
+                        index=np.arange(c_bottom_depths.iloc[index, 2]),
+                        columns=["l", "a", "b"],
                     )
                     lab_intpl_lyrs.append(lab_intpl)
 
                     # Create dummy data for lab_lyrs
-                    lab_dummy = [["", "", ""] for _ in range(len(hz_lyrs[i]))]
-                    lab_lyrs.append(dict(zip(hz_lyrs[i].keys(), lab_dummy)))
-
-                    # Empty osd_hz_lyrs entry
-                    osd_hz_lyrs.append("")
+                    lab_dummy = [["", "", ""] for _ in range(len(hz_lyrs[index]))]
+                    lab_lyrs.append(dict(zip(hz_lyrs[index].keys(), lab_dummy)))
 
                     # Create dummy data for munsell_lyrs
-                    munsell_dummy = [""] * len(hz_lyrs[i])
-                    munsell_lyrs.append(dict(zip(hz_lyrs[i].keys(), munsell_dummy)))
+                    munsell_dummy = [""] * len(hz_lyrs[index])
+                    munsell_lyrs.append(dict(zip(hz_lyrs[index].keys(), munsell_dummy)))
 
             # Series URL Generation
             # Initialize lists to store series URLs
@@ -2359,16 +1486,17 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             # Group data by 'cokey'
             OSDhorzdata_group_cokey = [g for _, g in OSDhorzdata_pd.groupby("cokey", sort=False)]
 
-            for group in OSDhorzdata_group_cokey:
+            for index, group in enumerate(OSDhorzdata_group_cokey):
                 # Check if compkind is not in OSD_compkind or if series contains any null values
                 if (
-                    mucompdata_pd.loc[i]["compkind"] not in OSD_compkind
+                    mucompdata_pd.loc[index]["compkind"] not in OSD_compkind
                     or group["series"].isnull().any()
                 ):
                     SDE_URL.append("")
                     SEE_URL.append("")
                 else:
-                    # Extract compname, convert to lowercase, remove trailing numbers, and replace spaces with underscores
+                    # Extract compname, convert to lowercase, remove trailing numbers, and replace
+                    # spaces with underscores
                     comp = group["compname"].iloc[0].lower()
                     comp = re.sub(r"\d+$", "", comp)
                     comp = comp.replace(" ", "_")
@@ -2381,7 +1509,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             # Initialize lists to store data layers and URLs
             lab_lyrs = []
             lab_intpl_lyrs = []
-            osd_hz_lyrs = []
             munsell_lyrs = []
             SDE_URL = []
             SEE_URL = []
@@ -2390,7 +1517,9 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             for i in range(len(mucompdata_pd)):
                 # Initialize a DataFrame filled with NaNs
                 lab_intpl = pd.DataFrame(
-                    np.nan, index=np.arange(c_bottom_depths.iloc[i, 2]), columns=["l", "a", "b"]
+                    np.nan,
+                    index=np.arange(c_bottom_depths.iloc[i, 2]),
+                    columns=["l", "a", "b"],
                 )
                 lab_intpl_lyrs.append(lab_intpl)
 
@@ -2401,7 +1530,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
                 # Append dummy data to lists
                 lab_lyrs.append(dict(zip(keys, lab_dummy)))
-                osd_hz_lyrs.append("")
                 munsell_lyrs.append(dict(zip(keys, munsell_dummy)))
 
                 # Append empty URLs
@@ -2412,7 +1540,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         # Initialize lists to store data layers and URLs
         lab_lyrs = []
         lab_intpl_lyrs = []
-        osd_hz_lyrs = []
         munsell_lyrs = []
         SDE_URL = []
         SEE_URL = []
@@ -2421,7 +1548,9 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         for i in range(len(mucompdata_pd)):
             # Initialize a DataFrame filled with NaNs
             lab_intpl = pd.DataFrame(
-                np.nan, index=np.arange(c_bottom_depths.iloc[i, 2]), columns=["l", "a", "b"]
+                np.nan,
+                index=np.arange(c_bottom_depths.iloc[i, 2]),
+                columns=["l", "a", "b"],
             )
             lab_intpl_lyrs.append(lab_intpl)
 
@@ -2432,7 +1561,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
             # Append dummy data to lists
             lab_lyrs.append(dict(zip(keys, lab_dummy)))
-            osd_hz_lyrs.append("")
             munsell_lyrs.append(dict(zip(keys, munsell_dummy)))
 
             # Append empty URLs
@@ -2497,7 +1625,8 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             munsell_lyrs,
         ) = layer_lists
 
-    # Create a new column 'soilID_rank' which will be True for the first row in each group sorted by 'distance' and False for other rows
+    # Create a new column 'soilID_rank' which will be True for the first row in each group sorted
+    # by 'distance' and False for other rows
     mucompdata_pd = mucompdata_pd.sort_values(["compname_grp", "distance"])
     mucompdata_pd["soilID_rank"] = ~mucompdata_pd.duplicated("compname_grp", keep="first")
 
@@ -2506,7 +1635,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
 
     mucompdata_pd = mucompdata_pd.reset_index(drop=True)
 
-    # -------------------------------------------------------------------------------------------------------------
+    # ---------------------------------------------------------------------------------------------
     # Ecological site
     if data_source == "SSURGO":
         if not out["ESD"]:
@@ -2520,7 +1649,10 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
             # Check if any cokey in ESD matches with mucompdata_pd
             if any(ESD["cokey"].isin(mucompdata_pd["cokey"])):
                 ESDcompdata_pd = pd.merge(
-                    mucompdata_pd[["mukey", "cokey", "compname"]], ESD, on="cokey", how="left"
+                    mucompdata_pd[["mukey", "cokey", "compname"]],
+                    ESD,
+                    on="cokey",
+                    how="left",
                 )
             else:
                 ESDcompdata_pd = None
@@ -2528,7 +1660,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     elif data_source == "STATSGO":
         ESDcompdataQry = (
             "SELECT cokey,  coecoclasskey, ecoclassname FROM coecoclass WHERE cokey IN ("
-            + ",".join(map(str, cokey_list))
+            + ",".join(map(str, comp_key))
             + ")"
         )
         ESDcompdata_out = sda_return(propQry=ESDcompdataQry)
@@ -2566,151 +1698,146 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         # Update ecoclassid based on MLRA update by querying 'ESD_class_synonym_list' table
         ecositeID = ESDcompdata_pd["ecoclassid"].dropna().tolist()
 
-        conn = getDataStore_Connection()
-        cur = conn.cursor()
-        sql1 = (
-            "SELECT lu.ES_ID, lu.Synonym FROM ESD_class_synonym_list AS lu WHERE lu.Synonym IN (%s)"
-            % ", ".join(map(str, ecositeID))
-        )
-        cur.execute(sql1)
-        results = cur.fetchall()
+        # old code
+        ESD_geo = []
+        ESD_geo.extend(ecositeID)
+        ESD_geo = [ESD_geo for ESD_geo in ESD_geo if str(ESD_geo) != "nan"]
+        ESD_geo = ESD_geo[0][1:5]
 
-        if not results:
-            ESDcompdata_pd["ecoclassid_update"] = ESDcompdata_pd["ecoclassid"]
-        else:
-            esd_lookup = pd.DataFrame(results, columns=["ecoclassid_update", "ecoclassid"])
-            ESDcompdata_pd = pd.merge(ESDcompdata_pd, esd_lookup, on="ecoclassid", how="left")
-            ESDcompdata_pd["ecoclassid_update"].fillna(ESDcompdata_pd["ecoclassid"], inplace=True)
-
-        ecositeID = ESDcompdata_pd["ecoclassid_update"].tolist()
-        ESD_geo = [ecosite[:4] for ecosite in ESDcompdata_pd["ecoclassid_update"].unique()]
-
-        # Filter out 'nan' from ESD_geo
-        ESD_geo = [geo for geo in ESD_geo if str(geo) != "nan"]
-
-        # Extract class URL data
-        class_url_list = []
-        for geo in ESD_geo:
-            class_url = (
-                f"https://edit.jornada.nmsu.edu/services/downloads/esd/{geo}/class-list.json"
-            )
-            try:
-                with urllib.request.urlopen(class_url, timeout=4) as response:
-                    class_data = json.load(response)
-                    class_url_list.append(
-                        pd.json_normalize(class_data["ecoclasses"])[["id", "legacyId"]]
-                    )
-            except:
-                pass
-
-        # Concatenate all the class URL data into a single dataframe
-        ESD_list_pd = (
-            pd.concat(class_url_list)
-            if class_url_list
-            else pd.DataFrame(columns=["id", "legacyID"])
+        class_url = "https://edit.jornada.nmsu.edu/services/downloads/esd/%s/class-list.json" % (
+            ESD_geo
         )
 
-        # Extract ESD URL data
-        ESD_URL = []
-        for eco_id in ecositeID:
-            if eco_id in ESD_list_pd["id"].tolist() or eco_id in ESD_list_pd["legacyId"].tolist():
-                ecosite_edit_id = ESD_list_pd[
-                    ESD_list_pd.apply(lambda r: r.str.contains(eco_id, case=False).any(), axis=1)
-                ]["id"].values[0]
-                ES_URL_t = f"https://edit.jornada.nmsu.edu/catalogs/esd/{ecosite_edit_id[1:5]}/{ecosite_edit_id}"
-                ESD_URL.append(ES_URL_t)
-            else:
+        try:
+            response = requests.get(class_url, timeout=4)
+            response.raise_for_status()  # Raise an exception for any HTTP error
+
+            ESD_list = response.json()
+            ESD_list_pd = json_normalize(ESD_list["ecoclasses"])[["id", "legacyId"]]
+            ESD_URL = []
+
+            if isinstance(ESD_list, list):
                 ESD_URL.append("")
+            else:
+                for i in range(len(ecositeID)):
+                    if (
+                        ecositeID[i] in ESD_list_pd["id"].tolist()
+                        or ecositeID[i] in ESD_list_pd["legacyId"].tolist()
+                    ):
+                        ecosite_edit_id = ESD_list_pd[
+                            ESD_list_pd.apply(
+                                lambda r: r.str.contains(ecositeID[i], case=False).any(),
+                                axis=1,
+                            )
+                        ]["id"].values[0]
+                        ES_URL_t = "https://edit.jornada.nmsu.edu/catalogs/esd/%s/%s" % (
+                            ESD_geo,
+                            ecosite_edit_id,
+                        )
+                        ESD_URL.append(ES_URL_t)
+                    else:
+                        ESD_URL.append("")
 
-        ESDcompdata_pd["esd_url"] = ESD_URL
+            ESDcompdata_pd = ESDcompdata_pd.assign(esd_url=ESD_URL)
+        except requests.exceptions.RequestException as err:
+            ESDcompdata_pd["esd_url"] = pd.Series(np.repeat("", len(ecositeID))).values
+            print("An error occurred:", err)
 
-        # Assign missing ESD for components that have other instances with an assinged ESD
+        # Assign missing ESD for components that have other instances with an assigned ESD
         if ESDcompdata_pd is not None:
             if (
-                ESDcompdata_pd.ecoclassid_update.isnull().any()
+                ESDcompdata_pd.ecoclassid.isnull().any()
                 or ESDcompdata_pd.ecoclassname.isnull().any()
             ):
                 ESDcompdata_pd["compname_grp"] = ESDcompdata_pd.compname.str.replace(r"[0-9]+", "")
-
-                # Group by component name without numbers
                 ESDcompdata_pd_comp_grps = [
-                    group for _, group in ESDcompdata_pd.groupby("compname_grp", sort=False)
+                    g for _, g in ESDcompdata_pd.groupby(["compname_grp"], sort=False)
                 ]
-
                 ecoList_out = []
-                for group in ESDcompdata_pd_comp_grps:
-                    if (
-                        len(group) == 1
-                        or group.ecoclassid_update.isnull().all()
-                        or group.ecoclassname.isnull().all()
-                    ):
-                        ecoList_out.append(group)
+                for i in range(len(ESDcompdata_pd_comp_grps)):
+                    comp_grps_temp = ESDcompdata_pd_comp_grps[i]
+                    if len(comp_grps_temp) == 1:
+                        ecoList_out.append(comp_grps_temp)
                     elif (
-                        group.ecoclassid_update.isnull().any()
-                        and len(group.ecoclassid_update.dropna().unique()) == 1
-                    ) and (
-                        group.ecoclassname.isnull().any()
-                        and len(group.ecoclassname.dropna().unique()) == 1
+                        comp_grps_temp.ecoclassid.isnull().all()
+                        or comp_grps_temp.ecoclassname.isnull().all()
                     ):
-                        unique_ecoclassid = group.ecoclassid_update.dropna().unique()[0]
-                        unique_ecoclassname = group.ecoclassname.dropna().unique()[0]
-                        unique_url = next((url for url in group.esd_url.unique() if url), "")
-
-                        group["ecoclassid_update"] = unique_ecoclassid
-                        group["ecoclassname"] = unique_ecoclassname
-                        group["esd_url"] = unique_url
-
-                        ecoList_out.append(group)
+                        ecoList_out.append(comp_grps_temp)
+                    elif (
+                        comp_grps_temp.ecoclassid.isnull().any()
+                        and len(comp_grps_temp.ecoclassid.dropna().unique()) == 1
+                    ) and (
+                        comp_grps_temp.ecoclassname.isnull().any()
+                        and len(comp_grps_temp.ecoclassname.dropna().unique()) == 1
+                    ):
+                        comp_grps_temp["ecoclassid"] = pd.Series(
+                            np.tile(
+                                comp_grps_temp.ecoclassid.dropna().unique().tolist(),
+                                len(comp_grps_temp),
+                            )
+                        ).values
+                        comp_grps_temp["ecoclassname"] = pd.Series(
+                            np.tile(
+                                comp_grps_temp.ecoclassname.dropna().unique().tolist(),
+                                len(comp_grps_temp),
+                            )
+                        ).values
+                        url = comp_grps_temp.esd_url.unique().tolist()
+                        url = [x for x in url if x != ""]
+                        if not url:
+                            comp_grps_temp["esd_url"] = pd.Series(
+                                np.tile("", len(comp_grps_temp))
+                            ).values
+                        else:
+                            comp_grps_temp["esd_url"] = pd.Series(
+                                np.tile(url, len(comp_grps_temp))
+                            ).values
+                        ecoList_out.append(comp_grps_temp)
                     else:
-                        ecoList_out.append(group)
-
+                        ecoList_out.append(comp_grps_temp)
                 ESDcompdata_pd = pd.concat(ecoList_out)
 
-            ESDcompdata_pd = ESDcompdata_pd.drop_duplicates(subset="cokey", keep="first")
-
-            esd_comp_list = []
-            for _, group in ESDcompdata_pd.groupby("cokey", sort=False):
-                if group["ecoclassname"].isnull().values.any():
-                    esd_comp_list.append({"ecoclassid": "", "ecoclassname": "", "esd_url": ""})
-                else:
-                    esd_comp_list.append(
-                        {
-                            "ecoclassid": group["ecoclassid_update"].tolist()[0],
-                            "ecoclassname": group["ecoclassname"].tolist()[0],
-                            "esd_url": group["esd_url"].tolist()[0],
+        ESDcompdata_group_cokey = [g for _, g in ESDcompdata_pd.groupby(["cokey"], sort=False)]
+        esd_comp_list = []
+        for i in range(len(ESDcompdata_group_cokey)):
+            if ESDcompdata_group_cokey[i]["ecoclassname"].isnull().values.any():
+                esd_comp_list.append({"ESD": {"ecoclassid": "", "ecoclassname": "", "esd_url": ""}})
+            else:
+                esd_comp_list.append(
+                    {
+                        "ESD": {
+                            "ecoclassid": ESDcompdata_group_cokey[i]["ecoclassid"].tolist(),
+                            "ecoclassname": ESDcompdata_group_cokey[i]["ecoclassname"].tolist(),
+                            "esd_url": ESDcompdata_group_cokey[i]["esd_url"].tolist(),
                         }
-                    )
+                    }
+                )
+        else:
+            esd_comp_list = []
+            for i in range(len(mucompdata_pd)):
+                esd_comp_list.append({"ESD": {"ecoclassid": "", "ecoclassname": "", "esd_url": ""}})
 
-    else:
-        # Initialize the ecosite data list
-        esd_comp_list = [
-            {"ecoclassid": "", "ecoclassname": "", "esd_url": ""} for _ in range(len(mucompdata_pd))
-        ]
+        # Add ecosite data to mucompdata_pd for testing output. In cases with multiple ESDs per
+        # component, only take the first.
+        if ESDcompdata_pd is None:
+            mucompdata_pd["ecoclassid"] = pd.Series(np.tile(np.nan, len(mucompdata_pd))).values
+            mucompdata_pd["ecoclassname"] = pd.Series(np.tile(np.nan, len(mucompdata_pd))).values
+        else:
+            mucompdata_pd = pd.merge(
+                mucompdata_pd,
+                ESDcompdata_pd[["cokey", "ecoclassid", "ecoclassname"]]
+                .drop_duplicates("cokey", keep="first")
+                .reset_index(drop=True),
+                on="cokey",
+                how="left",
+            )
 
-    # Add ecosite data to mucompdata_pd for testing output.
-    # In cases with multiple ESDs per component, only take the first.
-    if ESDcompdata_pd is None:
-        mucompdata_pd["ecoclassid_update"] = np.nan
-        mucompdata_pd["ecoclassname"] = np.nan
-    else:
-        # Merge data with the first unique ecosite data per component
-        mucompdata_pd = mucompdata_pd.merge(
-            ESDcompdata_pd[["cokey", "ecoclassid_update", "ecoclassname"]].drop_duplicates(
-                "cokey", keep="first"
-            ),
-            on="cokey",
-            how="left",
-        )
-
-    # Note: Need to devise a way to improve the identification of bedrock or restrictive layers using both comp data and OSD.
-    #      There are many instances where the last one or two component horizons are missing texture data and the OSD bottom depth
-    #      is shallower than the component bottom depth, indicating a shallower profile. But the horizon designations are generic, e.g.,
-    #      H1, H2, H3; thus not allowing their use in indentifying bedrock.
-
-    # --------------------------------------------------------------------------------------------------------------------------
+    # --------------------------------------------------------------------------------------------
     # SoilIDList output
 
     # ------------------------------------------------------------
+    # Define model version
+    model_version = "3.0"
     # Create the soilIDRank_output list by combining the dataframes of various data sources
     soilIDRank_output = [
         pd.concat(
@@ -2810,7 +1937,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         cec_lyrs,
         ph_lyrs,
         ec_lyrs,
-        osd_hz_lyrs,
         lab_lyrs,
         munsell_lyrs,
     ]
@@ -2827,7 +1953,6 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
         cec_lyrs,
         ph_lyrs,
         ec_lyrs,
-        osd_hz_lyrs,
         lab_lyrs,
         munsell_lyrs,
     ) = reordered_lists
@@ -2874,17 +1999,21 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     # Writing out list of data needed for soilIDRank
     if plot_id is None:
         soilIDRank_output_pd.to_csv(
-            "%s/soilIDRank_ofile1.csv" % current_app.config["DATA_BACKEND"], index=None, header=True
+            "%s/soilIDRank_ofile1.csv" % current_app.config["DATA_BACKEND"],
+            index=None,
+            header=True,
         )
         mucompdata_cond_prob.to_csv(
-            "%s/soilIDRank_ofile2.csv" % current_app.config["DATA_BACKEND"], index=None, header=True
+            "%s/soilIDRank_ofile2.csv" % current_app.config["DATA_BACKEND"],
+            index=None,
+            header=True,
         )
     else:
         output_data = json.dumps(
             {
                 "metadata": {
                     "location": "us",
-                    "model": "v2",
+                    "model": "v3",
                     "unit_measure": {
                         "distance": "m",
                         "depth": "cm",
@@ -2910,7 +2039,7 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     return {
         "metadata": {
             "location": "us",
-            "model": "v2",
+            "model": "v3",
             "unit_measure": {
                 "distance": "m",
                 "depth": "cm",
@@ -2925,9 +2054,9 @@ def getSoilLocationBasedUS(lon, lat, plot_id):
     }
 
 
-#####################################################################################################
-#                                   rankPredictionUS                                                #
-#####################################################################################################
+##############################################################################################
+#                                   rankPredictionUS                                         #
+##############################################################################################
 def rankPredictionUS(
     lon,
     lat,
@@ -2941,10 +2070,12 @@ def rankPredictionUS(
     cracks,
     plot_id=None,
 ):
-    # TODO: Future testing to see if deltaE2000 values should be incorporated into site data
-    # use 'getColor_deltaE2000_OSD_pedon' and helper functions located in local_functions_SoilID_v3.py
-
-    # ------------------------------------------------------------------------------------------------
+    """
+    TODO: Future testing to see if deltaE2000 values should be incorporated
+    into site data use 'getColor_deltaE2000_OSD_pedon' and helper functions
+    located in local_functions_SoilID_v3.py
+    """
+    # ---------------------------------------------------------------------------------------
     # ------ Load in user data --------#
     # Initialize the DataFrame from the input data
     soil_df = pd.DataFrame(
@@ -2979,10 +2110,10 @@ def rankPredictionUS(
 
     # Filter out rows without valid horizon data
     relevant_columns = ["soilHorizon", "rfvDepth", "lab_Color"]
-    soil_df = soil_df.dropna(how="all", subset=relevant_columns)
+    soil_df_slice = soil_df.dropna(how="all", subset=relevant_columns)
 
-    if soil_df.empty:
-        soil_df = None
+    if soil_df_slice.empty:
+        soil_df_slice = None
 
     if soil_df_slice is not None:
         soil_df_slice = soil_df_slice.reset_index(drop=True)
@@ -3177,7 +2308,7 @@ def rankPredictionUS(
     else:
         text_completeness = "SoilID data entry for this site is complete."
 
-    # --------------------------------------------------------------------------------------------------------------------------------------
+    # -------------------------------------------------------------------------------------------
     # Load in component data from soilIDList
     if plot_id is None:
         # Reading from file
@@ -3191,8 +2322,8 @@ def rankPredictionUS(
         modelRun = load_model_output(plot_id)
         if modelRun:
             record_id = modelRun[0]
-            soilIDRank_output_pd = pd.read_csv(cStringIO.StringIO(modelRun[2]))
-            mucompdata_pd = pd.read_csv(cStringIO.StringIO(modelRun[3]))
+            soilIDRank_output_pd = pd.read_csv(io.StringIO(modelRun[2]))
+            mucompdata_pd = pd.read_csv(io.StringIO(modelRun[3]))
         else:
             print("Cannot find a plot with this ID")
 
@@ -3216,7 +2347,8 @@ def rankPredictionUS(
     mucompdata_pd["nirrcapcl"] = mucompdata_pd["nirrcapcl"].apply(trim_fraction)
     mucompdata_pd["irrcapcl"] = mucompdata_pd["irrcapcl"].apply(trim_fraction)
 
-    # Create soil depth DataFrame and subset component depths based on max user depth if no bedrock specified
+    # Create soil depth DataFrame and subset component depths based on max user
+    # depth if no bedrock specified
     c_bottom_depths = mucompdata_pd[["cokey", "compname", "c_very_bottom"]]
     c_bottom_depths.columns = ["cokey", "compname", "bottom_depth"]
     slices_of_soil = pd.concat([p_bottom_depth, c_bottom_depths], axis=0).reset_index(drop=True)
@@ -3227,15 +2359,12 @@ def rankPredictionUS(
 
     # Determine the maximum depth based on bedrock and user input
     if bedrock is None:
-        max_depth = min(p_bottom_depth.bottom_depth.values[0], 120)
+        max_depth = min(p_bottom_depth.bottom_depth.values[0], 200)
     else:
-        max_depth = 120
+        max_depth = 200
 
     # Truncate depths in slices_of_soil that exceed max_depth
     slices_of_soil.loc[slices_of_soil.bottom_depth > max_depth, "bottom_depth"] = max_depth
-
-    # Calculate the non-soil slices for each entry in slices_of_soil
-    slices_of_non_soil = max_depth - slices_of_soil.bottom_depth
 
     # Initialize the soil_matrix
     soil_matrix = pd.DataFrame(
@@ -3299,9 +2428,7 @@ def rankPredictionUS(
 
         # Subset depth intervals to match user measured intervals
         horz_vars = [p_hz_data]
-        horz_vars.extend(
-            [group.reset_index(drop=True).loc[pedon_slice_index] for group in horz_vars_group_cokey]
-        )
+        horz_vars.extend([group.reset_index(drop=True).loc[pedon_slice_index] for group in groups])
 
         # Calculate similarity for each depth slice
         dis_mat_list = []
@@ -3309,13 +2436,16 @@ def rankPredictionUS(
             i
         ) in (
             soil_matrix.index
-        ):  # i should be an index of p_hz_data depth slices, e.g. if user only enters 100-120cm data, then i = 100:120
+        ):  # i should be an index of p_hz_data depth slices, e.g. if user only enters
+            # 100-120cm data, then i = 100:120
             slice_data = [horz.loc[i] for horz in horz_vars]
             sliceT = pd.concat(slice_data, axis=1).T
             """
-            Not all depth slices have the same user recorded data. Here we filter out data columns with missing data and use that to subset the component data.
-            If certain components are missing lots of data and the filering results in less than 2 soil properties, than we filter out data columns with missing USER data
-            and components with missing data will later be assigned the max dissimilarity across all horizons
+            Not all depth slices have the same user recorded data. Here we filter out data columns
+            with missing data and use that to subset the component data.
+            If certain components are missing lots of data and the filering results in less than 2
+            soil properties, than we filter out data columns with missing USER data and components
+            with missing data will later be assigned the max dissimilarity across all horizons
             """
             # Filter columns based on available data
             if i < bedrock:
@@ -3391,7 +2521,7 @@ def rankPredictionUS(
     else:
         D_horz = None
 
-    ##---Site Data Similarity
+    # ---Site Data Similarity---
     def compute_site_similarity(
         p_slope, mucompdata, slices, additional_columns=None, feature_weight=None
     ):
@@ -3542,13 +2672,16 @@ def rankPredictionUS(
             rank_id += 1
 
     D_final["Rank_Data"] = Rank_Data
-
-    # Code options for production API/testing output
+    """
+    Code options for production API/testing output
     # ----------------------------------------------------------------
-    # #Data output for testing
-    # D_final_loc = pd.merge(D_final, mucompdata_pd[['compname', 'cokey', 'mukey', 'distance_score', 'distance_score_norm', 'clay', 'taxorder', 'taxsubgrp', 'OSD_text_int', 'OSD_rfv_int', 'data_source', 'Rank_Loc', 'majcompflag', 'comppct_r', 'distance', 'nirrcapcl', 'nirrcapscl',
-    #    'nirrcapunit', 'irrcapcl', 'irrcapscl', 'irrcapunit', 'ecoclassid_update', 'ecoclassname']], on='compname', how='left')
-
+    #Data output for testing
+    D_final_loc = pd.merge(D_final, mucompdata_pd[['compname', 'cokey', 'mukey',
+    'distance_score', 'distance_score_norm', 'clay', 'taxorder', 'taxsubgrp', 'OSD_text_int',
+    'OSD_rfv_int', 'data_source', 'Rank_Loc', 'majcompflag', 'comppct_r', 'distance',
+    'nirrcapcl', 'nirrcapscl', 'nirrcapunit', 'irrcapcl', 'irrcapscl', 'irrcapunit',
+    'ecoclassid_update', 'ecoclassname']], on='compname', how='left')
+    """
     # Refactoring the code for data output
 
     # Merge D_final with additional data from mucompdata_pd
@@ -3658,23 +2791,24 @@ def rankPredictionUS(
     ).reset_index(drop=True)
 
     # Uncomment code for testing output
+    """
     # ----------------------------------------------------------------
     # Data formatting for testing
 
     # # Update LCC_I column
-    # D_final_loc['LCC_I'] = np.where(
-    #     (D_final_loc['irrcapcl'] == 'nan') | (D_final_loc['irrcapscl'] == 'nan'),
-    #     None,
-    #     D_final_loc['irrcapcl'] + "-" + D_final_loc['irrcapscl']
-    # )
-    #
-    # # Update LCC_NI column
-    # D_final_loc['LCC_NI'] = np.where(
-    #     (D_final_loc['nirrcapcl'] == 'nan') | (D_final_loc['nirrcapscl'] == 'nan'),
-    #     None,
-    #     D_final_loc['nirrcapcl'] + "-" + D_final_loc['nirrcapscl']
-    # )
+    D_final_loc['LCC_I'] = np.where(
+        (D_final_loc['irrcapcl'] == 'nan') | (D_final_loc['irrcapscl'] == 'nan'),
+        None, D_final_loc['irrcapcl'] + "-" + D_final_loc['irrcapscl']
+    )
+
+    # Update LCC_NI column
+    D_final_loc['LCC_NI'] = np.where(
+        (D_final_loc['nirrcapcl'] == 'nan') | (D_final_loc['nirrcapscl'] == 'nan'),
+        None,
+        D_final_loc['nirrcapcl'] + "-" + D_final_loc['nirrcapscl']
+    )
     # ----------------------------------------------------------------
+    """
 
     # Replace NaN values in the specified columns with 0.0
     D_final_loc[
@@ -3737,399 +2871,15 @@ def rankPredictionUS(
 
     # If 'record_id' is provided, save the output data
     if record_id is not None:
-        model_version = 2
+        model_version = 3
         save_rank_output(record_id, model_version, json.dumps(output_data))
 
     return output_data
-
+    """
     # Data return for testing
-    # return(D_final_loc[['compname', 'compname_grp', 'Rank_Loc', 'distance_score_norm',  'Rank_Data', 'Score_Data', 'Rank_Data_Loc', 'Score_Data_Loc','ecoclassid_update', 'ecoclassname', 'LCC_I', 'LCC_NI', 'taxorder', 'taxsubgrp', 'majcompflag', 'comppct_r', 'distance']])
+    return(D_final_loc[['compname', 'compname_grp', 'Rank_Loc', 'distance_score_norm',
+    'Rank_Data', 'Score_Data', 'Rank_Data_Loc', 'Score_Data_Loc','ecoclassid_update',
+    'ecoclassname', 'LCC_I', 'LCC_NI', 'taxorder', 'taxsubgrp', 'majcompflag',
+    'comppct_r', 'distance']])
     # ----------------------------------------------------------------
-
-
-#####################################################################################################
-#                                          getSoilGridsGlobal                                       #
-#####################################################################################################
-def getSoilGridsGlobal(lon, lat, plot_id=None):
-    # -------------------------------------------------------------------------------------------
-
-    ########### SoilGrids250  ############################################
-    # Construct the SoilGrids API v2 URL
-    sg_api = f"https://rest.isric.org/soilgrids/v2.0/properties/query?lon={lon}&lat={lat}&property=cfvo&property=cec&property=clay&property=phh2o&property=sand&value=mean"
-
-    try:
-        # Make the API request using the requests library
-        response = requests.get(sg_api, timeout=160)
-        response.raise_for_status()  # Check for unsuccessful status codes
-        sg_out = response.json()
-
-    except requests.RequestException as e:
-        # Log the error and set the status to unavailable
-        if plot_id is not None:
-            save_soilgrids_output(plot_id, 1, json.dumps({"status": "unavailable"}))
-        sg_out = {"status": "unavailable"}
-
-    # Use the 'extract_values' function to extract specific keys from the JSON
-    top_depths = extract_values(sg_out, "top_depth")
-    bottom_depths = extract_values(sg_out, "bottom_depth")
-    values = extract_values(sg_out, "mean")
-    names = extract_values(sg_out, "name")
-
-    # Convert extracted lists to DataFrames
-    df_top_depth = pd.DataFrame(top_depths, columns=["top_depth"])
-    df_bottom_depth = pd.DataFrame(bottom_depths, columns=["bottom_depth"])
-    df_values = pd.DataFrame(values, columns=["value"])
-
-    # Repeat names for each 6 rows since there are 6 depths
-    df_names = pd.DataFrame(names * 6, columns=["prop"])
-
-    # Concatenate the DataFrames to form a single DataFrame
-    sg_data = pd.concat([df_names, df_top_depth, df_bottom_depth, df_values], axis=1)
-
-    # Pivot the data based on bottom depth and property name, setting values to be the 'value' column
-    sg_data_w = sg_data.pivot(index="bottom_depth", columns="prop", values="value")
-
-    # Add the top and bottom depth columns to the wide DataFrame
-    sg_data_w["hzdept_r"] = df_top_depth["top_depth"].head(6).tolist()
-    sg_data_w["hzdepb_r"] = df_bottom_depth["bottom_depth"].head(6).tolist()
-
-    # Reset the index for the resulting DataFrame
-    sg_data_w.reset_index(drop=True, inplace=True)
-
-    # Check if all values in the specified columns are NaN
-    if sg_data_w[["sand", "clay", "cfvo"]].isnull().all().all():
-        if plot_id is not None:
-            save_soilgrids_output(plot_id, 1, json.dumps({"status": "unavailable"}))
-        return {"status": "unavailable"}
-    else:
-        # Apply the factor to the specific columns
-        cols_to_multiply = ["sand", "clay", "cfvo", "phh2o", "cec"]
-        sg_data_w[cols_to_multiply] = sg_data_w[cols_to_multiply].multiply(0.1)
-        # Calculate silt and texture for the DataFrame
-        sg_data_w["silt"] = sg_data_w.apply(silt_calc, axis=1)
-        sg_data_w["texture"] = sg_data_w.apply(getTexture, axis=1)
-
-        # Fetch SG wRB Taxonomy
-        # Construct the API URL for fetching soil data
-        api_url = f"https://rest.isric.org/soilgrids/v2.0/classification/query?lon={lon}&lat={lat}&number_classes=3"
-
-        # Fetch data from the API
-        try:
-            with request.urlopen(api_url, timeout=6) as response:
-                sg_tax = json.load(response)
-        except Exception as e:
-            # Handle data fetch failure
-            if plot_id is not None:
-                # Assuming the function `save_soilgrids_output` exists elsewhere in the code
-                save_soilgrids_output(plot_id, 1, json.dumps({"status": "unavailable"}))
-            sg_tax = None
-
-        # If data was successfully fetched, process it
-        if sg_tax:
-            # Create DataFrame from fetched data
-            sg_tax_prob = pd.DataFrame(sg_tax["wrb_class_probability"], columns=["WRB_tax", "Prob"])
-
-            # Sort DataFrame by probability
-            sg_tax_prob.sort_values("Prob", ascending=False, inplace=True)
-
-            # Fetch descriptions for the soil classifications
-            # Assuming the function `getSG_descriptions` exists elsewhere in the code
-            WRB_Comp_Desc = getSG_descriptions(sg_tax_prob["WRB_tax"].tolist())
-
-            # Merge the descriptions with the data
-            TAXNWRB_pd = pd.merge(sg_tax_prob, WRB_Comp_Desc, on="WRB_tax", how="left")
-            TAXNWRB_pd.index = ["Rank1", "Rank2", "Rank3"]
-        else:
-            # If data wasn't fetched, create an unavailable status DataFrame
-            TAXNWRB_pd = pd.DataFrame(
-                {
-                    "WRB_tax": [""],
-                    "Prob": [""],
-                    "Description_en": [""],
-                    "Management_en": [""],
-                    "Description_es": [""],
-                    "Management_es": [""],
-                    "Description_ks": [""],
-                    "Management_ks": [""],
-                    "Description_fr": [""],
-                    "Management_fr": [""],
-                }
-            )
-
-        # Avoid repetitive code by creating a function to get data and aggregate
-        def get_and_agg(variable, sg_data_w, bottom, return_depth=False):
-            pd_int = getProfile_SG(sg_data_w, variable, c_bot=False)
-            if return_depth == True:
-                pd_lpks, lpks_depths = agg_data_layer(
-                    data=pd_int.var_pct_intpl, bottom=bottom, depth=True
-                )
-                return (pd_lpks.replace(np.nan, ""), lpks_depths)
-            else:
-                pd_lpks = agg_data_layer(data=pd_int.var_pct_intpl, bottom=bottom, depth=False)
-                return pd_lpks.replace(np.nan, "")
-
-        sand_pd_lpks, lpks_depths = get_and_agg("sand", sg_data_w, bottom, return_depth=True)
-
-        variables = ["clay", "cfvo", "phh2o", "cec"]
-        dataframes = {var: get_and_agg(var, sg_data_w, bottom) for var in variables}
-
-        clay_pd_lpks = dataframes["clay"]
-        rfv_pd_lpks = dataframes["cfvo"]
-        pH_pd_lpks = dataframes["phh2o"]
-        cec_pd_lpks = dataframes["cec"]
-
-        # Additional texture calculation
-        texture_pd = pd.DataFrame([sand_pd_lpks, clay_pd_lpks]).T
-        texture_pd.columns = ["sand", "clay"]
-        texture_pd["silt"] = texture_pd.apply(silt_calc, axis=1)
-        texture_pd_lpks = texture_pd.apply(getTexture, axis=1).replace([None], "")
-
-        # SoilGrids API call version 1
-        model_version = 1
-
-        # Define keys and corresponding values for the 'components' dictionary
-        component_keys = [
-            "compname",
-            "probability",
-            "descriptionEN",
-            "managementEN",
-            "descriptionES",
-            "managementES",
-            "descriptionKS",
-            "managementKS",
-            "descriptionFR",
-            "managementFR",
-        ]
-        component_values = [
-            TAXNWRB_pd.WRB_tax,
-            TAXNWRB_pd.Prob,
-            TAXNWRB_pd.Description_en,
-            TAXNWRB_pd.Management_en,
-            TAXNWRB_pd.Description_es,
-            TAXNWRB_pd.Management_es,
-            TAXNWRB_pd.Description_ks,
-            TAXNWRB_pd.Management_ks,
-            TAXNWRB_pd.Description_fr,
-            TAXNWRB_pd.Management_fr,
-        ]
-
-        # Create 'components' dictionary using a loop
-        components_dict = {
-            k: dict(zip(TAXNWRB_pd.index, v)) for k, v in zip(component_keys, component_values)
-        }
-
-        # Create remaining dictionaries for the SoilGrids dictionary
-        remaining_keys = ["bottom_depth", "texture", "sand", "clay", "rock_fragments", "ph", "cec"]
-        remaining_values = [
-            lpks_depths,
-            texture_pd_lpks,
-            sand_pd_lpks,
-            clay_pd_lpks,
-            rfv_pd_lpks,
-            pH_pd_lpks,
-            cec_pd_lpks,
-        ]
-
-        # Create the overall SoilGrids dictionary
-        SoilGrids = {
-            "components": components_dict,
-            "bottom_depth": dict(list(zip(lpks_depths.index, lpks_depths))),
-            "bedrock": bottom,
-        }
-        SoilGrids.update(
-            {k: dict(list(zip(v.index, v))) for k, v in zip(remaining_keys, remaining_values)}
-        )
-
-        # Define the metadata dictionary
-        metadata = {
-            "location": "global",
-            "model": "v1",
-            "unit_measure": {
-                "depth": "cm",
-                "cec": "cmol(c)/kg",
-                "clay": "%",
-                "rock_fragments": "cm3/100cm3",
-                "sand": "%",
-            },
-        }
-
-        # If a plot_id is provided, save the SoilGrids output
-        if plot_id is not None:
-            save_soilgrids_output(
-                plot_id, model_version, json.dumps({"metadata": metadata, "soilGrids": SoilGrids})
-            )
-
-        # Return the final result
-        return {"metadata": metadata, "soilGrids": SoilGrids}
-
-
-#####################################################################################################
-#                                          getSoilGridsUS                                           #
-#####################################################################################################
-def getSoilGridsUS(lon, lat, plot_id=None):
-    ########### SoilGrids250  ############################################
-    # Construct the SoilGrids API v2 URL
-    sg_api = f"https://rest.isric.org/soilgrids/v2.0/properties/query?lon={lon}&lat={lat}&property=cfvo&property=cec&property=clay&property=phh2o&property=sand&value=mean"
-
-    try:
-        # Make the API request using the requests library
-        response = requests.get(sg_api, timeout=60)
-        response.raise_for_status()  # Check for unsuccessful status codes
-        sg_out = response.json()
-
-    except requests.RequestException as e:
-        # Log the error and set the status to unavailable
-        if plot_id is not None:
-            save_soilgrids_output(plot_id, 1, json.dumps({"status": "unavailable"}))
-        sg_out = {"status": "unavailable"}
-
-    # Use the 'extract_values' function to extract specific keys from the JSON
-    top_depths = extract_values(sg_out, "top_depth")
-    bottom_depths = extract_values(sg_out, "bottom_depth")
-    values = extract_values(sg_out, "mean")
-    names = extract_values(sg_out, "name")
-
-    # Convert extracted lists to DataFrames
-    df_top_depth = pd.DataFrame(top_depths, columns=["top_depth"])
-    df_bottom_depth = pd.DataFrame(bottom_depths, columns=["bottom_depth"])
-    df_values = pd.DataFrame(values, columns=["value"])
-
-    # Repeat names for each 6 rows since there are 6 depths
-    df_names = pd.DataFrame(names * 6, columns=["prop"])
-
-    # Concatenate the DataFrames to form a single DataFrame
-    sg_data = pd.concat([df_names, df_top_depth, df_bottom_depth, df_values], axis=1)
-
-    # Pivot the data based on bottom depth and property name, setting values to be the 'value' column
-    sg_data_w = sg_data.pivot(index="bottom_depth", columns="prop", values="value")
-
-    # Add the top and bottom depth columns to the wide DataFrame
-    sg_data_w["hzdept_r"] = df_top_depth["top_depth"].head(6).tolist()
-    sg_data_w["hzdepb_r"] = df_bottom_depth["bottom_depth"].head(6).tolist()
-
-    # Reset the index for the resulting DataFrame
-    sg_data_w.reset_index(drop=True, inplace=True)
-
-    # Check if all values in the specified columns are NaN
-    if sg_data_w[["sand", "clay", "cfvo"]].isnull().all().all():
-        if plot_id is not None:
-            save_soilgrids_output(plot_id, 1, json.dumps({"status": "unavailable"}))
-        return {"status": "unavailable"}
-    else:
-        # Apply the factor to the specific columns
-        cols_to_multiply = ["sand", "clay", "cfvo", "phh2o", "cec"]
-        sg_data_w[cols_to_multiply] = sg_data_w[cols_to_multiply].multiply(0.1)
-
-        # Calculate silt and texture for the DataFrame
-        sg_data_w["silt"] = sg_data_w.apply(silt_calc, axis=1)
-        sg_data_w["texture"] = sg_data_w.apply(getTexture, axis=1)
-
-        # # Pull component descriptions
-        # ST_Comp_List = []
-        # ST_Comp_List = TAXOUSDA_pd["TAXOUSDA"].drop_duplicates(keep='first').tolist()
-        # ST_Comp_Desc = getST_descriptions(ST_Comp_List)
-        # ST_Comp_Desc = ST_Comp_Desc.rename(index={0:'Rank1', 1:'Rank2', 2:'Rank3'})
-        # TAXOUSDA_pd = pd.concat([TAXOUSDA_pd, ST_Comp_Desc], axis=1)
-
-        # Avoid repetitive code by creating a function to get data and aggregate
-        def get_and_agg(variable, sg_data_w, bottom, return_depth=False):
-            pd_int = getProfile_SG(sg_data_w, variable, c_bot=False)
-            if return_depth == True:
-                pd_lpks, lpks_depths = agg_data_layer(
-                    data=pd_int.var_pct_intpl, bottom=bottom, depth=True
-                )
-                return (pd_lpks.replace(np.nan, ""), lpks_depths)
-            else:
-                pd_lpks = agg_data_layer(data=pd_int.var_pct_intpl, bottom=bottom, depth=False)
-                return pd_lpks.replace(np.nan, "")
-
-        sand_pd_lpks, lpks_depths = get_and_agg("sand", sg_data_w, bottom, return_depth=True)
-
-        variables = ["clay", "cfvo", "phh2o", "cec"]
-        dataframes = {var: get_and_agg(var, sg_data_w, bottom) for var in variables}
-
-        clay_pd_lpks = dataframes["clay"]
-        rfv_pd_lpks = dataframes["cfvo"]
-        pH_pd_lpks = dataframes["phh2o"]
-        cec_pd_lpks = dataframes["cec"]
-
-        # Additional texture calculation
-        texture_pd = pd.DataFrame([sand_pd_lpks, clay_pd_lpks]).T
-        texture_pd.columns = ["sand", "clay"]
-        texture_pd["silt"] = texture_pd.apply(silt_calc, axis=1)
-        texture_pd_lpks = texture_pd.apply(getTexture, axis=1).replace([None], "")
-
-        # SoilGrids API call version 1
-        model_version = 1
-
-        # Define properties to be included in the SoilGrids dictionary
-        properties = ["texture", "sand", "clay", "rock_fragments", "ph", "cec"]
-        dataframes = {
-            "texture": texture_pd_lpks,
-            "sand": sand_pd_lpks,
-            "clay": clay_pd_lpks,
-            "rock_fragments": rfv_pd_lpks,
-            "ph": pH_pd_lpks,
-            "cec": cec_pd_lpks,
-        }
-
-        # Construct the nested dictionary using dictionary comprehension
-        SoilGrids = {
-            "bottom_depth": dict(lpks_depths.items()),
-            "bedrock": bottom,
-            **{prop: dict(dataframes[prop].items()) for prop in properties},
-        }
-        # Define the common metadata and SoilGrids data to be used in both saving and returning.
-        data = {
-            "metadata": {
-                "location": "us",
-                "model": "v1",
-                "unit_measure": {
-                    "depth": "cm",
-                    "cec": "cmol(c)/kg",
-                    "clay": "%",
-                    "rock_fragments": "cm3/100cm3",
-                    "sand": "%",
-                },
-            },
-            "soilGrids": SoilGrids,
-        }
-
-        # Save the data if plot_id is provided
-        if plot_id:
-            save_soilgrids_output(plot_id, model_version, json.dumps(data))
-
-        # Return the constructed data
-        return data
-
-    # -------------------------------------------------------------------------------------------
-    # # This code was used for SoilGrids v1
-
-    # def getST_descriptions(ST_Comp_List):
-    #     try:
-    #         conn = getDataStore_Connection()
-    #         cur = conn.cursor()
-    #         ST_Comp_List = [x.encode('UTF8') for x in ST_Comp_List]
-    #         sql = 'SELECT Suborder, Description_en, Management_en, Description_es, Management_es, Description_ks, Management_ks, Description_fr, Management_fr FROM soil_taxonomy_desc WHERE Suborder IN (' + ''.join(str(ST_Comp_List)[1:-1]) + ')'
-    #         cur.execute(sql)
-    #         results = cur.fetchall()
-    #         data = pd.DataFrame(list(results))
-    #         data.columns = ['Suborder', 'Description_en', 'Management_en', 'Description_es', 'Management_es', 'Description_ks', 'Management_ks', 'Description_fr', 'Management_fr']
-    #         return data
-    #     except Exception, err:
-    #         print err
-    #         return None
-    #     finally:
-    #         conn.close()
-
-    # # This function applies a cubic spline model to interpolate values at every 1cm for the SoilGrids data
-
-    # def cspline_soil_lpks(data):
-    #     xm=[0,5,15,30,60,100,199]
-    #     ym=data.loc[['M.sl1','M.sl2','M.sl3','M.sl4','M.sl5','M.sl6','M.sl7'],0].tolist()
-    #     x_int=np.arange(0, 200, 1)
-    #     cs = CubicSpline(xm,ym,bc_type='natural')
-    #     int_vals=cs(x_int)
-    #     data=pd.Series(int_vals).apply(pd.to_numeric)
-    #     return pd.DataFrame(data=data, columns=['intp_vals'])
-    # -------------------------------------------------------------------------------------------
+    """
